### 2020-10-17 #2759 @whedon
[PRE REVIEW]: climpred: Verification of weather and climate forecasts
**Submitting author:** @bradyrx (<a href="http://orcid.org/0000-0002-2309-8245">Riley Brady</a>)
**Repository:** <a href="https://www.github.com/pangeo-data/climpred" target ="_blank">https://www.github.com/pangeo-data/climpred</a>
**Version:** v2.1.1
