### 2020-10-17 6015695
remove unused CI (#504)
