+++ joss/paper.bib
@@ -264,7 +264,8 @@ key = ""
   volume={20},
   number={2},
   pages={130--141},
-  year={1963}
+  year={1963},
+  doi={10.1175/1520-0469(1963)020<0130:DNF>2.0.CO;2},
 }
 
 
