### 2020-10-19 #2759 @jedbrown
@danielskatz This looks cool, but I'm conflicted (lead author is also at CU Boulder).
