### 2020-10-19 #2759 @danielskatz
Sorry I missed that - I assume @openjournals/jose-eics will make a different editor invitation
