### 2020-10-20 #2759 @bradyrx
> @danielskatz This looks cool, but I'm conflicted (lead author is also at CU Boulder).
>
> PS: as is Kelly (in reviewer list above)

Hi Jed! Bummer I didn't meet you during my time here -- glad to see a fellow Buff here. Kelly is a good friend of mine, although we haven't collaborated directly. I crossed her name out above to avoid CU Boulder bias. Thanks!
