### 2020-10-20 #2759 @kyleniemeyer
@danielskatz I think you meant openjournals/joss-eics... 😄

@dhhagan could you instead edit this submission? Thanks!
