+++ climpred/metrics.py
@@ -1,6 +1,7 @@
 import warnings
 
 import numpy as np
+
 from scipy.stats import norm
 from xskillscore import (
     Contingency,
@@ -283,10 +284,10 @@ def _pearson_r(forecast, verif, dim=None, **metric_kwargs):
         +-----------------+-----------+
 
     See also:
-        * xskillscore.pearson_r
-        * xskillscore.pearson_r_p_value
-        * climpred.pearson_r_p_value
-        * climpred.pearson_r_eff_p_value
+        * :py:func:`~xskillscore.pearson_r`
+        * :py:func:`~xskillscore.pearson_r_p_value`
+        * :py:func:`~climpred.metrics._pearson_r_p_value`
+        * :py:func:`~climpred.metrics._pearson_r_eff_p_value`
     """
     with warnings.catch_warnings():
         warnings.simplefilter("ignore", category=RuntimeWarning)
@@ -333,10 +334,10 @@ def _pearson_r_p_value(forecast, verif, dim=None, **metric_kwargs):
         +-----------------+-----------+
 
     See also:
-        * xskillscore.pearson_r
-        * xskillscore.pearson_r_p_value
-        * climpred.pearson_r
-        * climpred.pearson_r_eff_p_value
+        * :py:func:`~xskillscore.pearson_r`
+        * :py:func:`~xskillscore.pearson_r_p_value`
+        * :py:func:`~climpred.metrics._pearson_r`
+        * :py:func:`~climpred.metrics._pearson_r_eff_p_value`
     """
     # p value returns a runtime error when working with NaNs, such as on a climate
     # model grid. We can avoid this annoying output by specifically suppressing
@@ -472,8 +473,8 @@ def _pearson_r_eff_p_value(forecast, verif, dim=None, **metric_kwargs):
         +-----------------+-----------+
 
     See Also:
-        * climpred.effective_sample_size
-        * climpred.spearman_r_eff_p_value
+        * :py:func:`~climpred.metrics._effective_sample_size`
+        * :py:func:`~climpred.metrics._spearman_r_eff_p_value`
 
     Reference:
         * Bretherton, Christopher S., et al. "The effective number of spatial degrees of
@@ -541,10 +542,10 @@ def _spearman_r(forecast, verif, dim=None, **metric_kwargs):
         +-----------------+-----------+
 
     See also:
-        * xskillscore.spearman_r
-        * xskillscore.spearman_r_p_value
-        * climpred.spearman_r_p_value
-        * climpred.spearman_r_eff_p_value
+        * :py:func:`~xskillscore.spearman_r`
+        * :py:func:`~xskillscore.spearman_r_p_value`
+        * :py:func:`~climpred.metrics._spearman_r_p_value`
+        * :py:func:`~climpred.metrics._spearman_r_eff_p_value`
     """
     with warnings.catch_warnings():
         warnings.simplefilter("ignore", category=RuntimeWarning)
@@ -591,10 +592,10 @@ def _spearman_r_p_value(forecast, verif, dim=None, **metric_kwargs):
         +-----------------+-----------+
 
     See also:
-        * xskillscore.spearman_r
-        * xskillscore.spearman_r_p_value
-        * climpred.spearman_r
-        * climpred.spearman_r_eff_p_value
+        * :py:func:`~xskillscore.spearman_r`
+        * :py:func:`~xskillscore.spearman_r_p_value`
+        * :py:func:`~climpred.metrics.spearman_r`
+        * :py:func:`~climpred.metrics.spearman_r_eff_p_value`
     """
     # p value returns a runtime error when working with NaNs, such as on a climate
[…]
