+++ climpred/metrics.py
@@ -270,7 +270,7 @@ def _pearson_r(forecast, verif, dim=None, **metric_kwargs):
         forecast (xarray object): Forecast.
         verif (xarray object): Verification data.
         dim (str): Dimension(s) to perform metric over.
-        metric_kwargs (dict): see xskillscore.pearson_r
+        metric_kwargs (dict): see :py:func:`~xskillscore.pearson_r`
 
     Details:
         +-----------------+-----------+
@@ -390,7 +390,7 @@ def _effective_sample_size(forecast, verif, dim=None, **metric_kwargs):
         forecast (xarray object): Forecast.
         verif (xarray object): Verification data.
         dim (str): Dimension(s) to perform metric over.
-        metric_kwargs (dict): see xskillscore.effective_sample_size
+        metric_kwargs (dict): see :py:func:`~xskillscore.effective_sample_size`
 
     Details:
         +-----------------+-----------------+
@@ -459,7 +459,7 @@ def _pearson_r_eff_p_value(forecast, verif, dim=None, **metric_kwargs):
         forecast (xarray object): Forecast.
         verif (xarray object): Verification data.
         dim (str): Dimension(s) to perform metric over.
-        metric_kwargs (dict): see xskillscore.pearson_r_eff_p_value
+        metric_kwargs (dict): see :py:func:`~xskillscore.pearson_r_eff_p_value`
 
     Details:
         +-----------------+-----------+
@@ -528,7 +528,7 @@ def _spearman_r(forecast, verif, dim=None, **metric_kwargs):
         forecast (xarray object): Forecast.
         verif (xarray object): Verification data.
         dim (str): Dimension(s) to perform metric over.
-        metric_kwargs (dict): see xskillscore.spearman_r
+        metric_kwargs (dict): see :py:func:`~xskillscore.spearman_r`
 
     Details:
         +-----------------+-----------+
@@ -578,7 +578,7 @@ def _spearman_r_p_value(forecast, verif, dim=None, **metric_kwargs):
         forecast (xarray object): Forecast.
         verif (xarray object): Verification data.
         dim (str): Dimension(s) to perform metric over.
-        metric_kwargs (dict): see xskillscore.spearman_r_p_value
+        metric_kwargs (dict): see :py:func:`~xskillscore.spearman_r_p_value`
 
     Details:
         +-----------------+-----------+
@@ -594,8 +594,8 @@ def _spearman_r_p_value(forecast, verif, dim=None, **metric_kwargs):
     See also:
         * :py:func:`~xskillscore.spearman_r`
         * :py:func:`~xskillscore.spearman_r_p_value`
-        * :py:func:`~climpred.metrics.spearman_r`
-        * :py:func:`~climpred.metrics.spearman_r_eff_p_value`
+        * :py:func:`~climpred.metrics._spearman_r`
+        * :py:func:`~climpred.metrics._spearman_r_eff_p_value`
     """
     # p value returns a runtime error when working with NaNs, such as on a climate
     # model grid. We can avoid this annoying output by specifically suppressing
@@ -653,7 +653,7 @@ def _spearman_r_eff_p_value(forecast, verif, dim=None, **metric_kwargs):
         forecast (xarray object): Forecast.
         verif (xarray object): Verification data.
         dim (str): Dimension(s) to perform metric over.
-        metric_kwargs (dict): see xskillscore.spearman_r_eff_p_value
+        metric_kwargs (dict): see :py:func:`~xskillscore.spearman_r_eff_p_value`
 
     Details:
         +-----------------+-----------+
@@ -719,7 +719,7 @@ def _mse(forecast, verif, dim=None, **metric_kwargs):
         forecast (xarray object): Forecast.
         verif (xarray object): Verification data.
         dim (str): Dimension(s) to perform metric over.
[…]
