### 2020-10-20 354510d
Use `pre-commit` GitHub action for consistency
