### 2020-10-21 #2759 @dhhagan
👋  @darothen @eviatarbach @sarats @neerajdhanraj: would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow [these](https://joss.readthedocs.io/en/latest/review_criteria.html) guidelines.
