+++ climpred/tests/test_repr.py
@@ -19,7 +19,10 @@ def test_repr_PM(PM_da_initialized_1d, PM_da_control_1d, display_style):
 
 @pytest.mark.parametrize("display_style", ("html", "text"))
 def test_repr_HC(
-    hind_ds_initialized_1d, hist_ds_uninitialized_1d, observations_ds_1d, display_style,
+    hind_ds_initialized_1d,
+    hist_ds_uninitialized_1d,
+    observations_ds_1d,
+    display_style,
 ):
     """Test html repr."""
     with xr.set_options(display_style=display_style):
