+++ climpred/tests/test_bootstrap.py
@@ -40,7 +40,11 @@ def test_bootstrap_PM_keep_lead_attrs(perfectModelEnsemble_initialized_control):
     """Test bootstrap_perfect_model works lazily."""
     pm = perfectModelEnsemble_initialized_control
     pm.get_initialized().lead.attrs["units"] = "years"
-    s = pm.bootstrap(iterations=ITERATIONS, comparison="m2c", metric="mse",)
+    s = pm.bootstrap(
+        iterations=ITERATIONS,
+        comparison="m2c",
+        metric="mse",
+    )
     assert "units" in s.lead.attrs
     assert s.lead.attrs["units"] == pm.get_initialized().lead.attrs["units"]
 
@@ -57,7 +61,10 @@ def test_bootstrap_PM_lazy_results(
     else:
         pm = pm.compute()
     s = pm.bootstrap(
-        iterations=ITERATIONS, comparison=comparison, metric="mse", dim=dim,
+        iterations=ITERATIONS,
+        comparison=comparison,
+        metric="mse",
+        dim=dim,
     )
     assert dask.is_dask_collection(s) == chunk
 
@@ -95,7 +102,10 @@ def test_bootstrap_hindcast_lazy(
 @pytest.mark.slow
 @pytest.mark.parametrize("resample_dim", ["member", "init"])
 def test_bootstrap_hindcast_resample_dim(
-    hind_da_initialized_1d, hist_da_uninitialized_1d, observations_da_1d, resample_dim,
+    hind_da_initialized_1d,
+    hist_da_uninitialized_1d,
+    observations_da_1d,
+    resample_dim,
 ):
     """Test bootstrap_hindcast when resampling member or init and alignment
     same_inits."""
@@ -268,7 +278,8 @@ def test_bootstrap_by_stacking_chunked(
     PM_ds_initialized_1d_ym_cftime, PM_ds_control_1d_ym_cftime
 ):
     res_chunked = _bootstrap_by_stacking(
-        PM_ds_initialized_1d_ym_cftime.chunk(), PM_ds_control_1d_ym_cftime.chunk(),
+        PM_ds_initialized_1d_ym_cftime.chunk(),
+        PM_ds_control_1d_ym_cftime.chunk(),
     )
     assert dask.is_dask_collection(res_chunked)
     res_chunked = res_chunked.compute()
@@ -304,7 +315,10 @@ def test_bootstrap_by_stacking_two_var_dataset(
 @pytest.mark.slow
 @pytest.mark.parametrize("alignment", VALID_ALIGNMENTS)
 def test_bootstrap_hindcast_alignment(
-    hind_da_initialized_1d, hist_da_uninitialized_1d, observations_da_1d, alignment,
+    hind_da_initialized_1d,
+    hist_da_uninitialized_1d,
+    observations_da_1d,
+    alignment,
 ):
     """Test bootstrap_hindcast for all alginments when resampling member."""
     bootstrap_hindcast(
