+++ docs/source/api/climpred.metrics._std_ratio.rst
@@ -1,6 +1,7 @@
 climpred.metrics.\_std\_ratio
 =============================
 
+
 .. currentmodule:: climpred.metrics
 
 .. autofunction:: _std_ratio
