### 2020-10-21 b1e30d1
try removing EOF fixer
