### 2020-10-21 #2759 @neerajdhanraj
Great, I got your point.

On Wed, Oct 21, 2020 at 11:03 PM David H Hagan <notifications@github.com>
wrote:

> Hi @neerajdhanraj <https://github.com/neerajdhanraj> After I identify
> another reviewer, I will officially start the review which will take place
> in a new issue on GitHub. There will be checkboxes you can check off as the
> review progresses and you will comment either directly in that issue or on
> the software's original repository. An example of what that process looks
> like can be found here
> <https://github.com/openjournals/joss-reviews/issues/1075>. I hope to
> have the reviewer list locked down within the next few days.
>
