### 2020-10-22 5419519
Format files and add `push` trigger
