+++ docs/source/api/climpred.metrics.Metric.rst
@@ -5,18 +5,12 @@ climpred.metrics.Metric
 
 .. autoclass:: Metric
 
-   
+
    .. automethod:: __init__
 
-   
+
    .. rubric:: Methods
 
    .. autosummary::
-   
-      ~Metric.__init__
-   
-   
 
-   
-   
-   
+      ~Metric.__init__
