+++ .github/workflows/lint.yml
@@ -1,6 +1,8 @@
 name: Lint
 
-on: pull_request_target
+on:
+  - pull_request_target
+  - push
 
 jobs:
   lint:  # Runs linting package checks for code styling.
