### 2020-10-26 #2781 @bradyrx
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10/f89qdf is OK
> - 10/ggtpks is OK
> - 10/gd7hfq is OK
> - 10/f782pf is OK
> - 10/gbpwpt is OK
> - 10/f4jjvf is OK
> - 10/dp65gs is OK
> - 10/f5w2xr is OK
> - 10/gdqdmw is OK
> - 10/gfxr7x is OK
> - 10/gf4wzh is OK
> - 10/f7r9st is OK
> - 10/gf48c9 is OK
> - 10/gdsnf8 is OK
> - 10/gg9ss8 is OK
> - 10/gc9wzk is OK
> - 10/dpsjbp is OK
> - 10/ggvcqv is OK
> - 10/ggkt9s is OK
> - 10/gfz6s5 is OK
> - 10/gdb424 is OK
> - 10/ggtbv2 is OK
> - 10/gddfcs is OK
>
> MISSING DOIs
>
> - 10.1201/9780203734636-38 may be a valid DOI for title: Deterministic nonperiodic flow
>
> INVALID DOIs
>
> - None
> ```

FYI, I fixed the missing DOI here: https://github.com/pangeo-data/climpred/pull/505. I'll implement any other changes requested from @darothen and @neerajdhanraj there as well (unless they're big enough to require separate PRs).
