### 2020-10-27 e5967b7
Update linting CI GH action (#507)
* Use `pre-commit` GitHub action for consistency

* Update pre-commit hooks and add black config file

* Use `pull_request` for the trigger

Co-authored-by: Riley Brady <riley.brady@colorado.edu>
