+++ climpred/bootstrap.py
@@ -33,8 +33,9 @@ from .utils import (
     get_metric_class,
     lead_units_equal_control_time_stride,
     rechunk_to_single_chunk_if_more_than_one_chunk_along_dim,
-    shift_cftime_singular,
+    shift_cftime_singular, 
 )
+
 
 
 def _resample(hind, resample_dim):
