### 2020-10-27 2c92b7c
Test against newly released py39
