### 2020-10-27 4ce5e3d
Enable intersphinx mapping (#506)
* Enable intersphinx mapping in docs
