+++ ci/requirements/minimum-tests.yml
@@ -12,6 +12,7 @@ dependencies:
   - coveralls
   - pytest
   - pytest-cov
+  - pytest-xdist
   - pip
   - pip:
     - pytest-lazy-fixture
