+++ .github/workflows/climpred_testing.yml
@@ -34,7 +34,7 @@ jobs:
         run: conda list
       - name: Run tests
         run: |
-          pytest --cov=climpred --cov-report=xml
+          pytest -n auto --cov=climpred --cov-report=xml
       - name: Upload coverage to codecov
         uses: codecov/codecov-action@v1.0.7
         with:
