+++ .github/workflows/climpred_testing.yml
@@ -10,6 +10,7 @@ jobs:
         run:
           shell: bash -l {0}
     strategy:
+      fail-fast: false
       matrix:
         python-version: [3.6, 3.7, 3.8]
     steps:
