### 2020-10-27 8b6ecc7
Use `-rf` flag to get a list of failed tests
