### 2020-10-28 39bc6bb
Attempt at fixing race condition
