+++ climpred/tests/conftest.py
@@ -10,26 +10,26 @@ from climpred.utils import convert_time_index
 CALENDAR = PM_CALENDAR_STR.strip("Datetime").lower()
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_ds3v_initialized_1d():
     """MPI Perfect-model-framework initialized timeseries xr.Dataset with three
     variables."""
     return load_dataset("MPI-PM-DP-1D").isel(area=1, period=-1)
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_ds_initialized_1d(PM_ds3v_initialized_1d):
     """MPI Perfect-model-framework initialized timeseries xr.Dataset."""
     return PM_ds3v_initialized_1d.drop_vars(["sos", "AMO"])
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_da_initialized_1d(PM_ds_initialized_1d):
     """MPI Perfect-model-framework initialized timeseries xr.DataArray."""
     return PM_ds_initialized_1d["tos"]
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_da_initialized_1d_lead0(PM_da_initialized_1d):
     """MPI Perfect-model-framework initialized timeseries xr.DataArray in lead-0
     framework."""
@@ -40,78 +40,78 @@ def PM_da_initialized_1d_lead0(PM_da_initialized_1d):
     return da
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_ds_initialized_3d_full():
     """MPI Perfect-model-framework initialized global maps xr.Dataset."""
     return load_dataset("MPI-PM-DP-3D")
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_da_initialized_3d_full(PM_ds_initialized_3d_full):
     """MPI Perfect-model-framework initialized global maps xr.Dataset."""
     return PM_ds_initialized_3d_full["tos"]
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_ds_initialized_3d(PM_ds_initialized_3d_full):
     """MPI Perfect-model-framework initialized maps xr.Dataset of subselected North
     Atlantic."""
     return PM_ds_initialized_3d_full.sel(x=slice(120, 130), y=slice(50, 60))
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_da_initialized_3d(PM_ds_initialized_3d):
     """MPI Perfect-model-framework initialized maps xr.DataArray of subselected North
     Atlantic."""
     return PM_ds_initialized_3d["tos"]
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_ds3v_control_1d():
     """To MPI Perfect-model-framework corresponding control timeseries xr.Dataset with
     three variables."""
     return load_dataset("MPI-control-1D").isel(area=1, period=-1)
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_ds_control_1d(PM_ds3v_control_1d):
     """To MPI Perfect-model-framework corresponding control timeseries xr.Dataset."""
     return PM_ds3v_control_1d.drop_vars(["sos", "AMO"])
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_da_control_1d(PM_ds_control_1d):
     """To MPI Perfect-model-framework corresponding control timeseries xr.DataArray."""
     return PM_ds_control_1d["tos"]
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def PM_ds_control_3d_full():
[…]
