### 2020-10-28 98f66fa
Enable --verbose flag
