### 2020-10-28 33bc27b
Trace fixture execution
