### 2020-10-28 acaf382
Update conda env file
