+++ ci/requirements/minimum-tests.yml
@@ -6,7 +6,7 @@ dependencies:
   - coveralls
   - dask
   - eofs
-  - esmpy
+  - esmpy=*=mpi*
   - esmtools>=1.1.3
   - ipython
   - matplotlib
