### 2020-11-02 3179be8
Use mamba + pytest-xdist to speed up CI testing (#515)
* Use mamba to speed up CI testing
