### 2020-11-04 6c0c120
linregress so y always predicant and x predictor
