+++ esmtools/stats.py
@@ -332,11 +332,7 @@ def linear_slope(x, y=None, dim="time", nan_policy="none"):
 
 
 @is_xarray([0, 1])
-<<<<<<< HEAD
 def linregress(x, y=None, dim="time", nan_policy="none"):
-=======
-def linregress(x, y=None, dim='time', nan_policy='none'):
->>>>>>> Revert "linregress so y always predicant and x predictor"
     """Vectorized applciation of ``scipy.stats.linregress``.
 
     .. note::
@@ -348,11 +344,10 @@ def linregress(x, y=None, dim='time', nan_policy='none'):
         the original units per day (e.g. SST per day).
 
     Args:
-        x (xarray object): Independent variable (predictor) for linear regression.
-            If ``y`` is ``None``, treat ``x`` as the dependent variable and remove
-            slope over ``dim``.
-        y (xarray object, optional): Dependent variable (predictand) for linear
-            regression. If ``None``, treat ``x`` as the predictand.
+        y (xarray object): Dependent variable (predictand) for linear
+            regression.
+        x (xarray object, optional): Independent variable (predictor) for linear
+            regression. If ``x`` is ``None`` remove slope over ``dim``.
         dim (str, optional): Dimension to apply linear regression over.
             Defaults to "time".
         nan_policy (str, optional): Policy to use when handling nans. Defaults to
@@ -371,11 +366,7 @@ def linregress(x, y=None, dim='time', nan_policy='none'):
 
     """
     if y is None:
-<<<<<<< HEAD
         has_dims(x, dim, "predictand (x)")
-=======
-        has_dims(x, dim, 'predictand (x)')
->>>>>>> Revert "linregress so y always predicant and x predictor"
         X, slope_factor = _convert_time_and_return_slope_factor(x[dim], dim)
         Y = x
     else:
