### 2020-11-04 9190955
linregress so y always predicant and x predictor
