### 2020-11-05 677b3c3
added stats accessors
