### 2020-11-13 cc7128c
fix vectorize bug
