+++ esmtools/tests/conftest.py
@@ -10,9 +10,9 @@ def gridded_da_float():
     # https://alysivji.github.io/pytest-fixures-with-function-arguments.html
     def _gen_data():
         data = np.random.rand(60, 3, 3)
-        da = xr.DataArray(data, dims=["time", "lat", "lon"])
+        da = xr.DataArray(data, dims=['time', 'lat', 'lon'])
         # Annual resolution time axis for 60 years.
-        da["time"] = np.arange(1900, 1960)
+        da['time'] = np.arange(1900, 1960)
         return da
 
     return _gen_data
@@ -46,9 +46,9 @@ def gridded_da_datetime():
     # https://alysivji.github.io/pytest-fixures-with-function-arguments.html
     def _gen_data():
         data = np.random.rand(60, 3, 3)
-        da = xr.DataArray(data, dims=["time", "lat", "lon"])
+        da = xr.DataArray(data, dims=['time', 'lat', 'lon'])
         # Monthly resolution time axis for 5 years.
-        da["time"] = np.arange("1990-01", "1995-01", dtype="datetime64[M]")
+        da['time'] = np.arange('1990-01', '1995-01', dtype='datetime64[M]')
         return da
 
     return _gen_data
@@ -61,9 +61,9 @@ def gridded_da_cftime():
     # https://alysivji.github.io/pytest-fixures-with-function-arguments.html
     def _gen_data():
         data = np.random.rand(60, 3, 3)
-        da = xr.DataArray(data, dims=["time", "lat", "lon"])
+        da = xr.DataArray(data, dims=['time', 'lat', 'lon'])
         # Monthly resolution time axis for 5 years.
-        da["time"] = xr.cftime_range(start="1990-01", freq="M", periods=da.time.size)
+        da['time'] = xr.cftime_range(start='1990-01', freq='M', periods=da.time.size)
         return da
 
     return _gen_data
@@ -73,8 +73,8 @@ def gridded_da_cftime():
 def gridded_ds_float(gridded_da_float):
     """Mock data of a gridded time series as an xarray Dataset with float time."""
     data = xr.Dataset()
-    data["foo"] = gridded_da_float()
-    data["bar"] = gridded_da_float()
+    data['foo'] = gridded_da_float()
+    data['bar'] = gridded_da_float()
     return data
 
 
@@ -82,8 +82,8 @@ def gridded_ds_float(gridded_da_float):
 def gridded_ds_datetime(gridded_da_datetime):
     """Mock data of a gridded time series as an xarray Dataset with numpy datetime."""
     data = xr.Dataset()
-    data["foo"] = gridded_da_datetime()
-    data["bar"] = gridded_da_datetime()
+    data['foo'] = gridded_da_datetime()
+    data['bar'] = gridded_da_datetime()
     return data
 
 
@@ -91,8 +91,8 @@ def gridded_ds_datetime(gridded_da_datetime):
 def gridded_ds_cftime(gridded_da_cftime):
     """Mock data of a gridded time series as an xarray Dataset with cftime."""
     data = xr.Dataset()
-    data["foo"] = gridded_da_cftime()
-    data["bar"] = gridded_da_cftime()
+    data['foo'] = gridded_da_cftime()
+    data['bar'] = gridded_da_cftime()
     return data
 
 
@@ -103,9 +103,9 @@ def ts_monthly_da():
     # https://alysivji.github.io/pytest-fixures-with-function-arguments.html
     def _gen_data():
         data = np.random.rand(60)
-        da = xr.DataArray(data, dims=["time"])
+        da = xr.DataArray(data, dims=['time'])
         # Monthly resolution time axis for 10 years.
[…]
