+++ docs/source/api/esmtools.stats.ACF.rst
@@ -1,6 +1,0 @@
-esmtools.stats.ACF
-==================
-
-.. currentmodule:: esmtools.stats
-
-.. autofunction:: ACF
