### 2020-11-13 ef246a4
Add Github Actions system from climpred and xskillscore
