+++ .github/workflows/lint.yml
@@ -1,0 +1,19 @@
+
+name: Lint
+
+on: pull_request
+
+jobs:
+  lint:  # Runs linting package checks for code styling.
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          ref: ${{github.event.pull_request.head.ref}}
+          repository: ${{github.event.pull_request.head.repo.full_name}}
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.6
+      - name: Lint via pre-commit checks
+        uses: pre-commit/action@v2.0.0
