+++ .travis.yml
@@ -1,43 +1,0 @@
-# Based on http://conda.pydata.org/docs/travis.html
-language: python
-sudo: false # use container based build
-notifications:
-    email: false
-
-python:
-  - "3.6"
-
-before_install:
-  - wget http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
-  - bash miniconda.sh -b -p $HOME/miniconda
-  - export PATH="$HOME/miniconda/bin:$PATH"
-  - hash -r
-  - conda config --set always_yes yes --set changeps1 no --set show_channel_urls true
-  - conda update -q conda
-  - conda info -a
-
-install:
-  - conda env create -n esmtools-dev --file ci/environment-dev-3.6.yml
-  - source activate esmtools-dev
-  - conda install pip
-  - conda list
-  - which conda
-  - which python
-  - pip install .
-  - python setup.py install --user
-  - pip install pytest-cov
-  - pip install coveralls
-  - pip install black
-  - pip install flake8
-
-script:
-  - coverage run --source esmtools -m py.test
-  - coverage report
-  - pushd docs
-  - nbstripout source/*.ipynb
-  - make html
-  - popd
-  - ./ci/run-linter.sh
-
-after_success:
-  - coveralls
