+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 repos:
 
   -   repo: https://github.com/pre-commit/pre-commit-hooks
-      rev: v2.2.3
+      rev: v3.3.0
       hooks:
       -   id: trailing-whitespace
       -   id: end-of-file-fixer
@@ -11,27 +11,20 @@ repos:
       -   id: no-commit-to-branch
       -   id: debug-statements
       -   id: check-merge-conflict
+      -   id: check-added-large-files
 
-  -   repo: https://github.com/ambv/black
-      rev: 19.10b0
+  -   repo: https://github.com/psf/black
+      rev: stable
       hooks:
       - id: black
-        args: ["--line-length", "88", "--skip-string-normalization"]
+        args: ["--skip-string-normalization"]
 
-  -   repo: https://github.com/pre-commit/pre-commit-hooks
-      rev: v2.4.0
+  -   repo: https://gitlab.com/pycqa/flake8
+      rev: 3.8.4
       hooks:
       - id: flake8
-        args: ["--max-line-length=88", "--exclude=__init__.py", "--ignore=W605,W503,F722,C901"]
 
   -   repo: https://github.com/pre-commit/mirrors-isort
-      rev: v4.3.21
+      rev: v5.6.4
       hooks:
       -   id: isort
-          args: ["-w", "88"]
-
-  -   repo: https://github.com/PyCQA/doc8
-      rev: 0.8.1rc2
-      hooks:
-      -   id: doc8
-          args: ["--max-line-length", "100", "--ignore-path", "docs/source/setting-up-data.rst", "--allow-long-titles"]
