+++ setup.cfg
@@ -8,6 +8,9 @@ max-line-length = 88
 max-complexity = 18
 select = B,C,E,F,W,T4,B9
 
+[black]
+line-length = 88
+
 [isort]
 known_first_party=esmtools
 multi_line_output=3
