### 2020-11-13 f73cb13
add matplotlib to reqs
