+++ .pre-commit-config.yaml
@@ -7,7 +7,6 @@ repos:
       -   id: end-of-file-fixer
       -   id: check-docstring-first
       -   id: check-yaml
-      -   id: double-quote-string-fixer
       -   id: no-commit-to-branch
       -   id: debug-statements
       -   id: check-merge-conflict
@@ -17,7 +16,6 @@ repos:
       rev: stable
       hooks:
       - id: black
-        args: ["--skip-string-normalization"]
 
   -   repo: https://gitlab.com/pycqa/flake8
       rev: 3.8.4
