+++ esmtools/carbon.py
@@ -30,8 +30,8 @@ def calculate_compatible_emissions(global_co2_flux, co2atm_forcing):
           https://doi.org/10/f44bbn.
 
     """
-    compatible_emissions = co2atm_forcing.diff('time') * 2.12 - global_co2_flux
-    compatible_emissions.name = 'compatible_emissions'
+    compatible_emissions = co2atm_forcing.diff("time") * 2.12 - global_co2_flux
+    compatible_emissions.name = "compatible_emissions"
     return compatible_emissions
 
 
@@ -76,9 +76,9 @@ def co2_sol(t, s):
         return ff
 
     ff = xr.apply_ufunc(
-        sol_calc, t, s, input_core_dims=[[], []], vectorize=True, dask='allowed'
+        sol_calc, t, s, input_core_dims=[[], []], vectorize=True, dask="allowed"
     )
-    ff.attrs['units'] = 'mol/kg/atm'
+    ff.attrs["units"] = "mol/kg/atm"
     return ff
 
 
@@ -89,21 +89,21 @@ def get_iam_emissions():
         iam_emissions (xr.object): emissions from the IAMs in PgC/yr.
     """
     ds = []
-    member = ['rcp26', 'rcp45', 'rcp85']
+    member = ["rcp26", "rcp45", "rcp85"]
     for r in member:
-        if r == 'rcp26':
-            r = 'rcp3pd'
+        if r == "rcp26":
+            r = "rcp3pd"
         r = r.upper()
-        link = f'http://www.pik-potsdam.de/~mmalte/rcps/data/{r}_EMISSIONS.xls'
-        e = pd.read_excel(link, sheet_name=f'{r}_EMISSIONS', skiprows=35, header=2)
+        link = f"http://www.pik-potsdam.de/~mmalte/rcps/data/{r}_EMISSIONS.xls"
+        e = pd.read_excel(link, sheet_name=f"{r}_EMISSIONS", skiprows=35, header=2)
         e = e.set_index(e.columns[0])
-        e.index.name = 'Year'
-        ds.append(e[['FossilCO2', 'OtherCO2']].to_xarray())
-    ds = xr.concat(ds, 'member', **CONCAT_KWARGS)
-    ds = ds.sel(Year=slice(1850, 2100)).rename({'Year': 'time'})
-    ds['member'] = member
-    ds['IAM_emissions'] = ds['FossilCO2'] + ds['OtherCO2']
-    return ds['IAM_emissions']
+        e.index.name = "Year"
+        ds.append(e[["FossilCO2", "OtherCO2"]].to_xarray())
+    ds = xr.concat(ds, "member", **CONCAT_KWARGS)
+    ds = ds.sel(Year=slice(1850, 2100)).rename({"Year": "time"})
+    ds["member"] = member
+    ds["IAM_emissions"] = ds["FossilCO2"] + ds["OtherCO2"]
+    return ds["IAM_emissions"]
 
 
 def plot_compatible_emissions(
@@ -136,56 +136,56 @@ def plot_compatible_emissions(
         fig, ax = plt.subplots(figsize=(10, 6))
     # hist
     alpha = 0.1
-    c = 'gray'
+    c = "gray"
     compatible_emissions.isel(member=0).sel(
         time=slice(None, 2005)
-    ).to_dataframe().unstack()['compatible_emissions'].plot(
+    ).to_dataframe().unstack()["compatible_emissions"].plot(
         ax=ax, legend=False, color=c, alpha=alpha
     )
     compatible_emissions.isel(member=0).sel(time=slice(None, 2005)).mean(
-        'initialization'
-    ).plot(ax=ax, color='w', lw=3)
+        "initialization"
[…]
