+++ esmtools/tests/test_stats_poly.py
@@ -5,7 +5,7 @@ from xarray.testing import assert_allclose
 
 from esmtools.stats import polyfit, rm_poly, rm_trend
 
-TIME_TYPES = ['datetime', 'cftime', 'float']
+TIME_TYPES = ["datetime", "cftime", "float"]
 
 
 def _np_rm_poly(x, y, order):
@@ -14,10 +14,10 @@ def _np_rm_poly(x, y, order):
     return y - fit
 
 
-@pytest.mark.parametrize('gridded', (True, False))
-@pytest.mark.parametrize('time_type', TIME_TYPES)
-@pytest.mark.parametrize('order', (1, 2, 3, 4))
-@pytest.mark.parametrize('single_arg', (True, False))
+@pytest.mark.parametrize("gridded", (True, False))
+@pytest.mark.parametrize("time_type", TIME_TYPES)
+@pytest.mark.parametrize("order", (1, 2, 3, 4))
+@pytest.mark.parametrize("single_arg", (True, False))
 def test_rm_poly_against_time(
     gridded_da_datetime,
     gridded_da_cftime,
@@ -28,8 +28,8 @@ def test_rm_poly_against_time(
     single_arg,
 ):
     """Tests that ``rm_poly`` works against a time index."""
-    data = eval(f'gridded_da_{time_type}')()
-    x = data['time']
+    data = eval(f"gridded_da_{time_type}")()
+    x = data["time"]
     y = data if gridded else data.isel(lat=0, lon=0)
     if single_arg:
         detrended = rm_poly(y, order=order)
@@ -38,9 +38,9 @@ def test_rm_poly_against_time(
     assert (detrended != y).all()
 
 
-@pytest.mark.parametrize('gridded', (True, False))
-@pytest.mark.parametrize('time_type', TIME_TYPES)
-@pytest.mark.parametrize('single_arg', (True, False))
+@pytest.mark.parametrize("gridded", (True, False))
+@pytest.mark.parametrize("time_type", TIME_TYPES)
+@pytest.mark.parametrize("single_arg", (True, False))
 def test_rm_poly_against_time_dask(
     gridded_da_datetime,
     gridded_da_cftime,
@@ -50,8 +50,8 @@ def test_rm_poly_against_time_dask(
     single_arg,
 ):
     """Tests that ``rm_poly`` works against a time index with dask arrays."""
-    data = eval(f'gridded_da_{time_type}')()
-    x = data['time']
+    data = eval(f"gridded_da_{time_type}")()
+    x = data["time"]
     y = data if gridded else data.isel(lat=0, lon=0)
     if single_arg:
         expected = rm_poly(y, order=2)
@@ -62,8 +62,8 @@ def test_rm_poly_against_time_dask(
     assert_allclose(expected, actual)
 
 
-@pytest.mark.parametrize('gridded', (True, False))
-@pytest.mark.parametrize('order', (1, 2, 3, 4))
+@pytest.mark.parametrize("gridded", (True, False))
+@pytest.mark.parametrize("order", (1, 2, 3, 4))
 def test_rm_poly_against_data(gridded_da_float, gridded, order):
     """Tests that ``rm_poly`` works against a single data series and another gridded
     data series."""
@@ -73,7 +73,7 @@ def test_rm_poly_against_data(gridded_da_float, gridded, order):
     assert (detrended != y).all()
 
 
-@pytest.mark.parametrize('gridded', (True, False))
+@pytest.mark.parametrize("gridded", (True, False))
 def test_rm_poly_against_data_dask(gridded_da_float, gridded):
     """Tests that ``rm_poly`` works against a single data series and another gridded
     data series with dask arrays."""
@@ -84,11 +84,11 @@ def test_rm_poly_against_data_dask(gridded_da_float, gridded):
[…]
