+++ esmtools/__init__.py
@@ -1,3 +1,4 @@
+# flake8: noqa
 from pkg_resources import DistributionNotFound, get_distribution
 
 from . import (
@@ -16,6 +17,6 @@ from .versioning.print_versions import show_versions
 
 try:
     __version__ = get_distribution(__name__).version
-except DistributionNotFound:
+except DistributionNotFound:  # pragma: no cover
     # package is not installed
-    pass
+    pass  # pragma: no cover
