+++ requirements.txt
@@ -3,6 +3,7 @@ matplotlib
 numpy
 scipy
 statsmodels
+toolz
 tqdm
 xarray>=0.16.0
 xskillscore
