+++ README.rst
@@ -2,26 +2,59 @@
 
 a toolbox for Earth System Model analysis
 
-.. image:: https://travis-ci.org/bradyrx/esmtools.svg?branch=master
-    :target: https://travis-ci.org/bradyrx/esmtools
+..
+    Table version of badges inspired by pySTEPS.
 
-.. image:: https://img.shields.io/pypi/v/esmtools.svg
-    :target: https://pypi.python.org/pypi/esmtools/
+.. list-table::
+    :stub-columns: 1
+    :widths: 10 90
 
-.. image:: https://img.shields.io/conda/vn/conda-forge/esmtools.svg
-    :target: https://anaconda.org/conda-forge/esmtools
-    :alt: Conda Version
+    * - docs
+      - |docs|
+    * - tests
+      - |requires| |codecov|
+    * - package
+      - |conda| |pypi|
+    * - license
+      - |license|
+    * - community
+      - |contributors| |downloads|
+    * - tutorials
+      - |gallery|
 
-.. image:: https://coveralls.io/repos/github/bradyrx/esmtools/badge.svg?branch=master
-    :target: https://coveralls.io/github/bradyrx/esmtools?branch=master
-
-.. image:: https://img.shields.io/readthedocs/esmtools/stable.svg?style=flat
+.. |docs| image:: https://img.shields.io/readthedocs/esmtools/stable.svg?style=flat
     :target: https://esmtools.readthedocs.io/en/stable/?badge=stable
     :alt: Documentation Status
 
-.. image:: https://img.shields.io/github/license/bradyrx/esmtools.svg
+.. |requires| image:: https://requires.io/github/bradyrx/esmtools/requirements.svg?branch=master
+     :target: https://requires.io/github/bradyrx/esmtools/requirements/?branch=master
+     :alt: Requirements Status
+
+.. |codecov| image:: https://codecov.io/gh/bradyrx/esmtools/branch/master/graph/badge.svg?token=e53kXaaOqS
+      :target: https://codecov.io/gh/bradyrx/esmtools
+
+.. |conda| image:: https://img.shields.io/conda/vn/conda-forge/esmtools.svg
+    :target: https://anaconda.org/conda-forge/esmtools
+    :alt: Conda Version
+
+.. |pypi| image:: https://img.shields.io/pypi/v/esmtools.svg
+   :target: https://pypi.python.org/pypi/esmtools/
+
+.. |license| image:: https://img.shields.io/github/license/bradyrx/esmtools.svg
     :alt: license
-    :target: LICENSE.txt
+    :target: ../../LICENSE.txt
+
+.. |contributors| image:: https://img.shields.io/github/contributors/bradyrx/esmtools
+    :alt: GitHub contributors
+    :target: https://github.com/bradyrx/esmtools/graphs/contributors
+
+.. |downloads| image:: https://img.shields.io/conda/dn/conda-forge/esmtools
+    :alt: Conda downloads
+    :target: https://anaconda.org/conda-forge/esmtools
+
+.. |gallery| image:: https://img.shields.io/badge/esmtools-example_gallery-ed7b0e.svg
+    :alt: climpred gallery
+    :target: https://esmtools.readthedocs.io/en/stable/examples.html
 
 Installation
 ============
