### 2020-11-13 27e5817
Add Github Actions system from climpred and xskillscore (#103)
* Add Github Actions system for CI
* Removes Travis CI
* Switches to CodeCov for coverage testing
* Fixes bug with vectorize=True with new dask/xarray
