+++ esmtools/tests/conftest.py
@@ -127,7 +127,12 @@ def ts_annual_da():
 @pytest.fixture()
 def annual_all_leap():
     """Mock annual 12-year time series with an all leap calendar."""
-    data = xr.DataArray(np.random.rand(12,), dims=["time"])
+    data = xr.DataArray(
+        np.random.rand(
+            12,
+        ),
+        dims=["time"],
+    )
     data["time"] = xr.cftime_range(
         "1990", freq="YS", periods=data.time.size, calendar="all_leap"
     )
@@ -137,7 +142,12 @@ def annual_all_leap():
 @pytest.fixture()
 def annual_no_leap():
     """Mock annual 12-year time series with a no leap calendar."""
-    data = xr.DataArray(np.random.rand(12,), dims=["time"])
+    data = xr.DataArray(
+        np.random.rand(
+            12,
+        ),
+        dims=["time"],
+    )
     data["time"] = xr.cftime_range(
         "1990", freq="YS", periods=data.time.size, calendar="noleap"
     )
@@ -147,7 +157,12 @@ def annual_no_leap():
 @pytest.fixture()
 def annual_gregorian():
     """Mock annual 12-year time series with a Gregorian calendar."""
-    data = xr.DataArray(np.random.rand(12,), dims=["time"])
+    data = xr.DataArray(
+        np.random.rand(
+            12,
+        ),
+        dims=["time"],
+    )
     data["time"] = xr.cftime_range(
         "1990", freq="YS", periods=data.time.size, calendar="gregorian"
     )
@@ -157,7 +172,12 @@ def annual_gregorian():
 @pytest.fixture()
 def annual_julian():
     """Mock annual 12-year time series with a Julian calendar."""
-    data = xr.DataArray(np.random.rand(12,), dims=["time"])
+    data = xr.DataArray(
+        np.random.rand(
+            12,
+        ),
+        dims=["time"],
+    )
     data["time"] = xr.cftime_range(
         "1990", freq="YS", periods=data.time.size, calendar="julian"
     )
