+++ esmtools/stats.py
@@ -26,8 +26,8 @@ def _check_y_not_independent_variable(y, dim):
     """
     if isinstance(y, xr.DataArray) and (y.name == dim):
         raise ValueError(
-            f'Dependent variable y should not be the same as the dim {dim} being '
-            'applied over. Change your y variable to x.'
+            f"Dependent variable y should not be the same as the dim {dim} being "
+            "applied over. Change your y variable to x."
         )
 
 
@@ -85,12 +85,12 @@ def _handle_nans(x, y, nan_policy):
     # Only support 1D, since we are doing `~np.isnan()` indexing for 'omit'/'drop'.
     if (x.ndim > 1) or (y.ndim > 1):
         raise ValueError(
-            f'x and y must be 1-dimensional. Got {x.ndim} for x and {y.ndim} for y.'
+            f"x and y must be 1-dimensional. Got {x.ndim} for x and {y.ndim} for y."
         )
 
-    if nan_policy in ['none', 'propagate']:
+    if nan_policy in ["none", "propagate"]:
         return x, y
-    elif nan_policy == 'raise':
+    elif nan_policy == "raise":
         if has_missing(x) or has_missing(y):
             raise ValueError(
                 "Input data contains NaNs. Consider changing `nan_policy` to 'none' "
@@ -98,7 +98,7 @@ def _handle_nans(x, y, nan_policy):
             )
         else:
             return x, y
-    elif nan_policy in ['omit', 'drop']:
+    elif nan_policy in ["omit", "drop"]:
         if has_missing(x) or has_missing(y):
             x_mod, y_mod = match_nans(x, y)
             # The above function pairwise-matches nans. Now we remove them so that we
@@ -138,11 +138,11 @@ def _polyfit(x, y, order, nan_policy):
     """
     x_mod, y_mod = _handle_nans(x, y, nan_policy)
     # This catches cases where a given grid cell is full of nans, like in land masking.
-    if (nan_policy in ['omit', 'drop']) and (x_mod.size == 0):
+    if (nan_policy in ["omit", "drop"]) and (x_mod.size == 0):
         return np.full(len(x), np.nan)
     # This catches cases where there is missing values in the independent axis, which
     # breaks polyfit.
-    elif (nan_policy in ['none', 'propagate']) and (has_missing(x_mod)):
+    elif (nan_policy in ["none", "propagate"]) and (has_missing(x_mod)):
         return np.full(len(x), np.nan)
     else:
         # fit to data without nans, return applied to original independent axis.
@@ -161,14 +161,14 @@ def _warn_if_not_converted_to_original_time_units(x):
         if x.timeutils.is_temporal:
             if x.timeutils.freq is None:
                 warnings.warn(
-                    'Datetime frequency not detected. Slope and std. errors will be '
-                    'in original units per day (e.g., degC per day). Multiply by '
-                    'e.g., 365.25 to convert to original units per year.'
+                    "Datetime frequency not detected. Slope and std. errors will be "
+                    "in original units per day (e.g., degC per day). Multiply by "
+                    "e.g., 365.25 to convert to original units per year."
                 )
 
 
 @is_xarray(0)
-def autocorr(ds, dim='time', nlags=None):
+def autocorr(ds, dim="time", nlags=None):
     """Compute the autocorrelation function of a time series to a specific lag.
 
     .. note::
@@ -200,12 +200,12 @@ def autocorr(ds, dim='time', nlags=None):
[…]
