### 2020-11-14 1339ed5
implement fix predictability horizon
