### 2020-11-15 fdf2d25
ref_compute_kwargs now taking dim and metric
