### 2020-11-17 f1f0f09
update miniconda version in CI
