### 2020-11-17 63743cb
update second instance of miniconda install
