### 2020-11-17 ca2e747
set reuse_weights to False in testing
