+++ climpred/tests/test_smoothing.py
@@ -64,7 +64,9 @@ def test_spatial_smoothing_xesmf_reduce_spatial_dims_MPI_curv(
     step = 5
     # RXB: For some reason, `reuse_weights` breaks here with newer xesmf/esmpy version.
     # ValueError: to reuse weights, you need to provide either filename or weights
-    actual = spatial_smoothing_xesmf(da, d_lon_lat_kws={"lon": step}, reuse_weights=False)
+    actual = spatial_smoothing_xesmf(
+        da, d_lon_lat_kws={"lon": step}, reuse_weights=False
+    )
     expected_lat_size = 180 // step
     assert actual["lon"].size < da.lon.size
     assert actual["lat"].size == expected_lat_size
@@ -79,7 +81,9 @@ def test_spatial_smoothing_xesmf_reduce_spatial_dims_CESM(
     step = 0.1
     # RXB: For some reason, `reuse_weights` breaks here with newer xesmf/esmpy version.
     # ValueError: to reuse weights, you need to provide either filename or weights
-    actual = spatial_smoothing_xesmf(da, d_lon_lat_kws={"lat": step}, reuse_weights=False)
+    actual = spatial_smoothing_xesmf(
+        da, d_lon_lat_kws={"lat": step}, reuse_weights=False
+    )
     # test whether upsampled
     assert actual["lon"].size >= da.nlon.size
     assert actual["lat"].size >= da.nlat.size
@@ -107,8 +111,12 @@ def test_compute_after_smooth_goddard_2013(
     """Test compute_perfect_model works after smoothings."""
     # RXB: For some reason, `reuse_weights` breaks here with newer xesmf/esmpy version.
     # ValueError: to reuse weights, you need to provide either filename or weights
-    PM_da_control_3d_full = smooth_goddard_2013(PM_da_control_3d_full, reuse_weights=False)
-    PM_da_initialized_3d_full = smooth_goddard_2013(PM_da_initialized_3d_full, reuse_weights=False)
+    PM_da_control_3d_full = smooth_goddard_2013(
+        PM_da_control_3d_full, reuse_weights=False
+    )
+    PM_da_initialized_3d_full = smooth_goddard_2013(
+        PM_da_initialized_3d_full, reuse_weights=False
+    )
     actual = compute_perfect_model(PM_da_initialized_3d_full, PM_da_control_3d_full)
     north_atlantic = actual.sel(lat=slice(40, 50), lon=slice(-30, -20))
     assert not north_atlantic.isnull().any()
