### 2020-11-17 03767f7
update climpred dev requirements
