### 2020-11-18 e170b45
linregress so y always predicant and x predictor
