+++ esmtools/stats.py
@@ -344,11 +344,10 @@ def linregress(x, y=None, dim="time", nan_policy="none"):
         the original units per day (e.g. SST per day).
 
     Args:
-        x (xarray object): Independent variable (predictor) for linear regression.
-            If ``y`` is ``None``, treat ``x`` as the dependent variable and remove
-            slope over ``dim``.
-        y (xarray object, optional): Dependent variable (predictand) for linear
-            regression. If ``None``, treat ``x`` as the predictand.
+        y (xarray object): Dependent variable (predictand) for linear
+            regression.
+        x (xarray object, optional): Independent variable (predictor) for linear
+            regression. If ``x`` is ``None`` remove slope over ``dim``.
         dim (str, optional): Dimension to apply linear regression over.
             Defaults to "time".
         nan_policy (str, optional): Policy to use when handling nans. Defaults to
