### 2020-11-18 493a344
added stats accessors
