### 2020-11-21 febde3c
better docstrings and 3d test
