### 2020-11-25 20c005d
Stats func accessor 59 (#99)
* Adds accessors for stats functions.
