+++ climpred/tests/test_smoothing.py
@@ -62,10 +62,9 @@ def test_spatial_smoothing_xesmf_reduce_spatial_dims_MPI_curv(
     """Test whether spatial dimsizes are properly reduced."""
     da = PM_da_control_3d_full
     step = 5
-    # RXB: For some reason, `reuse_weights` breaks here with newer xesmf/esmpy version.
-    # ValueError: to reuse weights, you need to provide either filename or weights
     actual = spatial_smoothing_xesmf(
-        da, d_lon_lat_kws={"lon": step}, reuse_weights=False
+        da,
+        d_lon_lat_kws={"lon": step},
     )
     expected_lat_size = 180 // step
     assert actual["lon"].size < da.lon.size
@@ -79,10 +78,9 @@ def test_spatial_smoothing_xesmf_reduce_spatial_dims_CESM(
     """Test whether spatial dimsizes are properly reduced."""
     da = reconstruction_ds_3d_full
     step = 0.1
-    # RXB: For some reason, `reuse_weights` breaks here with newer xesmf/esmpy version.
-    # ValueError: to reuse weights, you need to provide either filename or weights
     actual = spatial_smoothing_xesmf(
-        da, d_lon_lat_kws={"lat": step}, reuse_weights=False
+        da,
+        d_lon_lat_kws={"lat": step},
     )
     # test whether upsampled
     assert actual["lon"].size >= da.nlon.size
@@ -93,9 +91,9 @@ def test_smooth_goddard_2013(PM_da_control_3d_full):
     """Test whether Goddard 2013 recommendations are fulfilled by
     smooth_Goddard_2013."""
     da = PM_da_control_3d_full
-    # RXB: For some reason, `reuse_weights` breaks here with newer xesmf/esmpy version.
-    # ValueError: to reuse weights, you need to provide either filename or weights
-    actual = smooth_goddard_2013(da, reuse_weights=False)
+    actual = smooth_goddard_2013(
+        da,
+    )
     # test that x, y not in dims
     assert "x" not in actual.dims
     assert "y" not in actual.dims
@@ -109,13 +107,11 @@ def test_compute_after_smooth_goddard_2013(
     PM_da_initialized_3d_full, PM_da_control_3d_full
 ):
     """Test compute_perfect_model works after smoothings."""
-    # RXB: For some reason, `reuse_weights` breaks here with newer xesmf/esmpy version.
-    # ValueError: to reuse weights, you need to provide either filename or weights
     PM_da_control_3d_full = smooth_goddard_2013(
-        PM_da_control_3d_full, reuse_weights=False
+        PM_da_control_3d_full,
     )
     PM_da_initialized_3d_full = smooth_goddard_2013(
-        PM_da_initialized_3d_full, reuse_weights=False
+        PM_da_initialized_3d_full,
     )
     actual = compute_perfect_model(PM_da_initialized_3d_full, PM_da_control_3d_full)
     north_atlantic = actual.sel(lat=slice(40, 50), lon=slice(-30, -20))
@@ -176,9 +172,7 @@ def test_HindcastEnsemble_spatial_smoothing_dim_and_skill(hindcast_recon_3d, dim
     """Test that HindcastEnsemble.smooth({dim: int}) aggregates dim."""
     he = hindcast_recon_3d
     smooth_kws = {key: step for key in dim}
-    # RXB: For some reason, `reuse_weights` breaks here with newer xesmf/esmpy version.
-    # ValueError: to reuse weights, you need to provide either filename or weights
-    he_smoothed = he.smooth(smooth_kws, reuse_weights=False)
+    he_smoothed = he.smooth(smooth_kws)
     for d in dim:
         assert he_smoothed.get_initialized()[d].any()
         assert he_smoothed.get_observations()[d].any()
