### 2020-11-25 bb3634d
update miniconda version in CI (#524)
* update miniconda version in CI

* set reuse_weights to False as default

* update climpred dev requirements
