### 2020-11-25 9d0d407
fix dim in verify reference, use dim_ori deepcopied dim
