+++ README.rst
@@ -1,6 +1,5 @@
 .. image:: https://i.imgur.com/HPOdOsR.png
 
-
 Verification of weather and climate forecasts.
 
 ..
