+++ climpred/tests/conftest.py
@@ -280,9 +280,10 @@ def hindcast_recon_3d(hind_ds_initialized_3d, reconstruction_ds_3d):
 
 @pytest.fixture()
 def hindcast_recon_1d_ym(hind_ds_initialized_1d, reconstruction_ds_1d):
-    """HindcastEnsemble initialized with `initialized`, `uninitialzed` and `recon`."""
-    hindcast = HindcastEnsemble(hind_ds_initialized_1d)
-    hindcast = hindcast.add_observations(reconstruction_ds_1d)
+    """HindcastEnsemble initialized with `initialized` and `recon`."""
+    hindcast = HindcastEnsemble(hind_ds_initialized_1d).add_observations(
+        reconstruction_ds_1d
+    )
     hindcast = hindcast - hindcast.sel(time=slice("1964", "2014")).mean("time").sel(
         init=slice("1964", "2014")
     ).mean("init")
