### 2020-12-08 68a00d8
document bugfix #522 in CL
