+++ climpred/__init__.py
@@ -8,8 +8,8 @@ from . import (
     constants,
     exceptions,
     graphics,
+    horizon,
     metrics,
-    predictability_horizon,
     prediction,
     relative_entropy,
     smoothing,
