### 2020-12-08 52a12ab
fix smoothing test
