+++ climpred/tests/test_smoothing.py
@@ -228,7 +228,7 @@ def test_HindcastEnsemble_smooth_carries_lead_attrs(hindcast_recon_1d_ym):
     he = hindcast_recon_1d_ym
     he_smooth = he.smooth({"lead": 4}, how="mean")
     assert (
-        he_smooth.verify(metric="rmse", comparison="m2e", dim="init").lead.attrs[
+        he_smooth.verify(metric="rmse", comparison="e2o", dim="init").lead.attrs[
             "units"
         ]
         == "years"
