### 2020-12-08 fb96620
fix alignment in smooth test
