+++ climpred/tests/test_smoothing.py
@@ -229,7 +229,7 @@ def test_HindcastEnsemble_smooth_carries_lead_attrs(hindcast_recon_1d_ym):
     he_smooth = he.smooth({"lead": 4}, how="mean")
     assert (
         he_smooth.verify(
-            metric="rmse", comparison="e2o", dim="init", alginment="same_verifs"
+            metric="rmse", comparison="e2o", dim="init", alignment="same_verifs"
         ).lead.attrs["units"]
         == "years"
     )
