+++ climpred/graphics.py
@@ -277,7 +277,8 @@ def plot_lead_timeseries_hindcast(
     if variable is None:
         variable = list(he.get_initialized().data_vars)[0]
     hind = he.get_initialized()[variable]
-    lead_freq = get_lead_cftime_shift_args(hind.lead.attrs["units"], 1)[1]
+    lead_freq = get_lead_cftime_shift_args(hind.lead.attrs["units"], 1)
+    lead_freq = str(lead_freq[0])+lead_freq[1]
     hist = he.get_uninitialized()
     if isinstance(hist, xr.Dataset):
         hist = hist[variable]
