### 2021-01-07 0c1be64
fix plot for composite lead units (#533)
* fix plot

* add lead attrs units to label

* shorten lead attrs unit in label
