+++ CHANGELOG.rst
@@ -2,6 +2,15 @@
 What's New
 ==========
 
+
+Bug Fixes
+---------
+
+- :py:meth:`~climpred.classes.PredictionEnsemble.plot` now shifts composite lead
+  frequencies like ``days``, ``pentads``, ``seasons`` correctly.
+  (:issue:`532`, :pr:`533`) `Aaron Spring`_.
+
+
 climpred v2.1.1 (2020-10-13)
 ============================
 
