### 2021-01-11 fdec3af
Reimplement predictability horizon (#521)
* implement fix predictability horizon

* ref_compute_kwargs now taking dim and metric

* better docstrings and 3d test

* fix dim in verify reference, use dim_ori deepcopied dim

* fix carrying lead #527

* document bugfix #522 in CL

* rename horizon

* fix smoothing test

* Rename test_perfectModelEnsemble_class.py to test_PerfectModelEnsemble_class.py

* Rename test_hindcastEnsemble_class.py to test_HindcastEnsemble_class.py
