+++ docs/source/api/climpred.horizon.horizon.rst
@@ -1,0 +1,6 @@
+climpred.horizon.horizon
+========================
+
+.. currentmodule:: climpred.horizon
+
+.. autofunction:: horizon
