### 2021-01-13 b18dfcd
Add temporary step for numba on py3.9
