### 2021-01-22 8e38f29
Update climpred_installs.yml
