+++ .github/workflows/climpred_installs.yml
@@ -1,6 +1,12 @@
 name: climpred installs
 
-on: pull_request
+on:
+  push:
+    branches:
+      - '*'
+  pull_request:
+    branches:
+      - '*'
 
 jobs:
   install-climpred:  # Installs climpred on various OS.
