### 2021-01-22 24d655f
Update climpred_testing.yml
