+++ .github/workflows/climpred_testing.yml
@@ -1,6 +1,12 @@
 name: climpred testing
 
-on: pull_request
+on:
+  push:
+    branches:
+      - '*'
+  pull_request:
+    branches:
+      - '*'
 
 jobs:
   test:  # Runs testing suite on various python versions.
