+++ .github/workflows/lint.yml
@@ -1,6 +1,12 @@
 name: Lint
 
-on: pull_request
+on:
+  push:
+    branches:
+      - '*'
+  pull_request:
+    branches:
+      - '*'
 
 jobs:
   lint:  # Runs linting package checks for code styling.
