### 2021-01-22 8200528
Update CHANGELOG.rst
