+++ CHANGELOG.rst
@@ -2,7 +2,6 @@
 What's New
 ==========
 
-
 climpred v2.1.2 (2021-xx-xx)
 ============================
 
