### 2021-01-22 a8958e4
Remove secrets (#535)
* Update climpred_testing.yml

* Update climpred_installs.yml

* Update lint.yml
