+++ .github/workflows/climpred_testing.yml
@@ -11,9 +11,6 @@ jobs:
         python-version: [3.6, 3.7, 3.8]
     steps:
       - uses: actions/checkout@v2
-        with:
-          ref: ${{github.event.pull_request.head.ref}}
-          repository: ${{github.event.pull_request.head.repo.full_name}}
       - name: Install Conda environment
         uses: conda-incubator/setup-miniconda@v1
         with:
@@ -43,9 +40,6 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-        with:
-          ref: ${{github.event.pull_request.head.ref}}
-          repository: ${{github.event.pull_request.head.repo.full_name}}
       - name: Install Conda environment
         uses: conda-incubator/setup-miniconda@v1
         with:
