### 2021-01-22 2a8ab87
Fix stable secrets (#538)
* Update climpred_installs.yml

* Update climpred_testing.yml

* Update lint.yml
