### 2021-01-22 8c03374
Update climpred_testing.yml (#539)
* Update climpred_testing.yml

* Update climpred_installs.yml
