### 2021-01-22 0deef0d
Update climpred_testing.yml
