### 2021-01-23 2f42937
Update minimum-tests.yml
