### 2021-01-30 7a9cfd7
Test against newly released py39 (#514)
* Test against newly released py39

* Add temporary step for numba on py3.9

* Use numba/label/dev channel for the time being

* Update climpred_testing.yml

* Update climpred_installs.yml

* Test against newly released py39

* Add temporary step for numba on py3.9

* Use numba/label/dev channel for the time being

* Update climpred_testing.yml

* Update climpred_installs.yml

* Remove temporary fix

* Revert to using temporary fix

Co-authored-by: Aaron Spring <aaronspring@users.noreply.github.com>
