### 2021-02-08 #2781 @aaronspring
Installing MurCCS for performance comparison is quite difficult. I couldn't get MurCCS running locally with py27 and basemap dependencies, but it is pre-installed on our supercomputer where I did these performance comparisons. I don't see a way to verify these numbers other trusting me on these or rely on Chris Kadow @ckadow, a MurCCS developer how did the MurCCS timings and I discussed these results with in Hamburg.
