### 2021-02-09 #2781 @bradyrx
@aaronspring I believe you have shown me some notebooks in the past with different timings, was there a gist of that that could be cleaned up? Otherwise, I agree completely with what you're saying.
