### 2021-02-09 c8bf8fe
Update examples.rst
