### 2021-02-09 06d012b
Update examples.rst
