+++ docs/source/examples.rst
@@ -2,7 +2,7 @@ Examples
 #########
 
 .. note::
-    Please use the `climpred-dev` enviroment (`conda install -f ci/requirements/climpred-dev.yml`)
+    Please use the `climpred-dev` enviroment (`conda env create -f ci/requirements/climpred-dev.yml`)
     to be able to complete all example notebooks listed here.
 
 Dask
