### 2021-02-09 #2781 @dhhagan
@samjsilva91 @aaronspring @bradyrx I think there are two simple solutions:

  1. Add a link to the notebook containing the performance comparisons that can serve as a reference with details on the type of cluster/HPC that was used
  2. Remove/soften the wording around the performance comparison..?

I will also ping some other editors and see if they have suggestions. Let me know if either of these approaches work for everyone.
