### 2021-02-09 #2781 @ckadow
Not sure, if I can help with this. If Aaron sends me the results of
MurCSS (and some ID), I could confirm these!?

As MurCSS development stopped, we appreciate the authors effort to
establish a new software around this topic, and kindly serve as a
reference.

Best regards,
Chris
