### 2021-02-09 #2781 @aaronspring
Verification of the NCAR 40member DPLE tas variable without bootstraping for 3 leads (1,2-5,6-9) and 3 metrics (murcss default) correlation/ACC, conditional bias and MSSS:
climpred: 8 sec
![by default 2021-02-09 at 16 05 26](https://user-images.githubusercontent.com/12237157/107382962-c5504700-6af0-11eb-8d53-1a7bae002040.png)
see gist: https://gist.github.com/aaronspring/57ab1ced1086106636aa87490943eab5
previously: https://gist.github.com/aaronspring/6a2bc517a1295d5e947d3d9408b37ecb

MurCCS: 438s = 8min
![image](https://user-images.githubusercontent.com/12237157/107376344-13158100-6aea-11eb-8b37-ff42aedca195.png)
Source is password-protected but @ckadow can confirm: https://www-miklip.dkrz.de/history/900515/results/

Therefore we wrote: "This leads to verification of a global 5° x 5° resolution climate prediction in a few seconds, compared to the 8 minutes required by MurCSS." We should change to 8 seconds.

I want to add that such benchmarking is always biased: http://matthewrocklin.com/blog/work/2017/03/09/biased-benchmarks, because pipelines are different. MurCCS also caches preprocessed data, climpred stores preprocessed files in zarr/netcdf here.

Plotting doesnt add much time: stays below 8s
![image](https://user-images.githubusercontent.com/12237157/107383933-c7ff6c00-6af1-11eb-8cc1-0214d3bc2c14.png)

What we wanted to show is that climpred enables an interactive workflow, which these numbers justify in our opinion.
