### 2021-02-09 #2781 @ckadow
Hello,

I confirm that the results are done by MurCSS at the DKRZ HPC and it
took the time shown.

However, I agree that this "time" comparison is misleading and comparing
apples and oranges. MurCSS does the whole chain at once including the
preprocessing directly from the CMIP/ESGF data standard (80% of the
time), which needs to be done in ClimPred as well (the ~10 min the
authors state in preprocess ->
https://gist.github.com/aaronspring/57ab1ced1086106636aa87490943eab5).
Of course, if you redo the same experiments afterwards, it is much
faster then. But if you change something in the chain, which is done in
the preprocessing (like adding a new reference model experiment), this
brings you back as well in ClimPred. MurCSS was always about user
convenience - one command does it all.  ClimPred on the other-side gives
much more control in the steps in between and can develop new views and
results on the fly, which is great and flexible. That should be
highlighted.

Nevertheless, ClimPred is faster in the skill score calculation and it
should be mentioned! Please just don't compare 8 minutes to 8 seconds.
Otherwise, readers expect a magic box.

Best regards,
Chris
