+++ requirements.txt
@@ -3,10 +3,12 @@ eofs
 esmtools>=1.1.3
 ipython
 matplotlib
-numpy
+numpy<=1.19
 pandas
 scipy
 toolz
 xarray>=0.16.1
 xrft
 xskillscore>=0.0.18
+netcdf4
+nc-time-axis
