+++ docs/source/examples.rst
@@ -1,6 +1,9 @@
 Examples
 #########
 
+.. note::
+    Please use the `climpred-dev` enviroment (`conda env create -f ci/requirements/climpred-dev.yml`)
+    to be able to complete all example notebooks listed here.
 
 Dask
 ====
