### 2021-02-09 #2781 @aaronspring
I agree we should frame this performance claim in the perspective of flexibility, where some part of performance increase comes from. We will work on new sentences. In the end we both use the same hardware. So for sure the timings delta comes from usability design and frameworks used. Our focus for climpred is wall time for the user and we expect that a user will work repeatedly with the same datasets. Here our modular design and flexibility is our strength (good for research exploring some detail of predictive skill in the same experiments over and over), whereas MurCCS needs to repeat everything every time (good for a Skript running on new hindcasts).

We didn’t agree on the goal: I agree we are not comparing apples and apples, but apples and oranges. My question is: “How fast can a user work repeatedly with the software?”

when we go to the 1000 bootstrapping iterations this initial preprocessing matters less, but we didn’t use this example in the draft.

the other part of performance increase comes from having the data in memory, whereas murcss has repeated IO, because it is file based with cdo calls, which makes it very resilient.
