### 2021-02-11 47a5019
Update examples.rst
