### 2021-02-11 15b09a2
Update examples.rst
