+++ docs/source/examples.rst
@@ -5,7 +5,7 @@ Examples
     Please use the ``climpred-dev`` enviroment
 
     .. code:: bash
-        (`conda env create -f ci/requirements/climpred-dev.yml`)
+        conda env create -f ci/requirements/climpred-dev.yml
 
     to be able to complete all example notebooks listed here.
 
