### 2021-02-11 883b5e0
Update examples.rst
