+++ docs/source/examples.rst
@@ -4,10 +4,10 @@ Examples
 .. note::
     Please use the ``climpred-dev`` enviroment
 
-    .. code:: bash
         conda env create -f ci/requirements/climpred-dev.yml
 
-    to be able to complete all example notebooks listed here.
+    to ensure that all dependencies are installed to complete all example
+    notebooks listed here.
 
 Dask
 ====
