### 2021-02-11 #2781 @ckadow
hi Aaron,

I agree, sounds good with the suggestions you mention. Exactly, if you
go towards bootstrapping, speed will be an important thing.

Best,
Chris
