### 2021-02-11 #2781 @aaronspring
great to here @samjsilva91 - we are currently working on this PR https://github.com/pangeo-data/climpred/pull/505 on the changes in the manuscript, mainly the performance claim (discussion above). we'll ping you once we merged this.
