### 2021-02-11 20f12e1
Update examples.rst
