### 2021-02-11 b7d8201
Update examples.rst and notebooks with dependency comments (#546)
* Update examples.rst
