+++ climpred/bootstrap.py
@@ -204,8 +204,8 @@ def _distribution_to_ci(ds, ci_low, ci_high, dim="iteration"):
     """
     ds = rechunk_to_single_chunk_if_more_than_one_chunk_along_dim(ds, dim)
     for v in ds.data_vars:
-      if np.issubdtype(ds[v].dtype, np.bool_):
-        ds[v] = ds[v].astype(np.float_)  # fails on py>36 if boolean dtype
+        if np.issubdtype(ds[v].dtype, np.bool_):
+            ds[v] = ds[v].astype(np.float_)  # fails on py>36 if boolean dtype
     return ds.quantile(q=[ci_low, ci_high], dim=dim, skipna=False)
 
 
