+++ climpred/bootstrap.py
@@ -208,8 +208,8 @@ def _distribution_to_ci(ds, ci_low, ci_high, dim="iteration"):
             if np.issubdtype(ds[v].dtype, np.bool_):
                 ds[v] = ds[v].astype(np.float_)  # fails on py>36 if boolean dtype
     else:
-        if np.issubdtype(ds[v].dtype, np.bool_):
-            ds[v] = ds[v].astype(np.float_)  # fails on py>36 if boolean dtype
+        if np.issubdtype(ds.dtype, np.bool_):
+            ds = ds.astype(np.float_)  # fails on py>36 if boolean dtype
     return ds.quantile(q=[ci_low, ci_high], dim=dim, skipna=False)
 
 
