### 2021-02-14 16654e4
Update minimum-tests.yml
