+++ climpred/classes.py
@@ -472,7 +472,9 @@ class PredictionEnsemble:
                 tsmooth_kws = None
                 for c in self._datasets['initialized'].coords:
                     if c not in CLIMPRED_DIMS:
-                        self._datasets['initialized'].attrs.update(self._datasets['observations'].attrs)
+                        self._datasets["initialized"].attrs.update(
+                            self._datasets["observations"].attrs
+                        )
                         
             elif "lead" in smooth_kws or "time" in smooth_kws:
                 smooth_fct = temporal_smoothing
