### 2021-02-14 c0650b6
unfix numpy version (#547)
* unfix jumpy version

* Update bootstrap.py
