### 2021-02-15 e0c23c0
Update smoothing.py
try with ds_out as xr.Dataset with lon and lat
