### 2021-02-15 1cd4a03
Update smoothing.py
