+++ joss/paper.bib
@@ -219,15 +219,6 @@ key = ""
   number = {8},
 }
 
-@inproceedings{Kluyver:2016,
-  title = {Jupyter Notebooks - a Publishing Format for Reproducible Computational Workflows},
-  booktitle = ELPUB,
-  author = {Kluyver, Thomas and {Ragan-Kelley}, Benjamin and P{\'e}rez, Fernando and Granger, Brian E. and Bussonnier, Matthias and Frederic, Jonathan and Kelley, Kyle and Hamrick, Jessica B. and Grout, Jason and Corlay, Sylvain and Ivanov, Paul and Avila, Dami{\'a}n and Abdalla, Safia and Willing, Carol and {al}, et},
-  year = {2016},
-  doi = {10/gf48c9},
-  abstract = {It is increasingly necessary for researchers in all fields to write computer code, and in order to reproduce research results, it is important that this code is published. We present Jupyter notebooks, a document format for publishing code, results and explanations in a form that is both readable and executable. We discuss various tools and use cases for notebook documents.},
-}
-
 @article{Kroger:2017,
   title = {Full-Field Initialized Decadal Predictions with the {{MPI}} Earth System Model: An Initial Shock in the {{North Atlantic}}},
   shorttitle = {Full-Field Initialized Decadal Predictions with the {{MPI}} Earth System Model},
