### 2021-02-16 ecb6634
make list more readable
