+++ joss/paper.md
@@ -59,14 +59,19 @@ are generally organized in our companion package, `xskillscore`.
 While other climate verification packages exist (_e.g._, `s2dverification`
 [@Manubens:2018] written in R and `MurCSS` [@Illing:2014] written with python-based
 `CDO`-bindings [@CDO]), `climpred` is unique for many reasons.
+
 1. `climpred` spans broad temporal scales of prediction, supporting the weather,
 subseasonal-to-seasonal (S2S), and seasonal-to-decadal (S2D) communities.
+
 2. `climpred` is highly modular and supports the research process from end-to-end,
 from loading in model output, to interactive pre-processing and analysis, to
 visualization.
+
 3. `climpred` supports `dask` [@dask; @Rocklin:2015] and thus works across all
 computational scales, from personal laptops to supercomputers (HPC).
+
 4. Flexibility and scaling leads to verification of global 5° x 5° resolution climate predictions in 8 seconds, compared to the 8 minutes required by `MurCSS`. However, note that `climpred` modularizes its workflow such that the verification step is performed on already pre-processed output, while `MurCSS` uses a more rigid framework that always required pre-processing. This time scale of seconds allows for a truly interactive analysis experience.
+
 5. `climpred` is part of the wider scientific python community, `pangeo`
 [@Eynard:2019]. A wide adoption of `climpred` could standardize prediction model
 evaluation and make verification reproducible [@Irving:2015].
