### 2021-02-16 fc1dcab
JOSS Review (#505)
* Edits to paper for JOSS
