### 2021-02-16 e6cbedd
Update climpred/tests/conftest.py
