### 2021-02-16 1506264
unfix xesmf  (#548)
* Update minimum-tests.yml

* Update classes.py

* Update climpred/tests/conftest.py
