### 2021-02-16 8f44f2d
Update paper.md (#550)
* Update paper.md
