+++ joss/paper.md
@@ -70,7 +70,12 @@ visualization.
 3. `climpred` supports `dask` [@dask; @Rocklin:2015] and thus works across all
 computational scales, from personal laptops to supercomputers (HPC).
 
-4. Flexibility and scaling leads to verification of global 5° x 5° resolution climate predictions in 8 seconds, compared to the 8 minutes required by `MurCSS`. However, note that `climpred` modularizes its workflow such that the verification step is performed on already pre-processed output, while `MurCSS` uses a more rigid framework that always required pre-processing. This time scale of seconds allows for a truly interactive analysis experience.
+4. Flexibility and scaling leads to verification of global 5° x 5° resolution climate
+predictions in 8 seconds, compared to the 8 minutes required by `MurCSS`. However,
+note that `climpred` modularizes its workflow such that the verification step is
+performed on already pre-processed output, while `MurCSS` uses a more rigid framework
+that always required pre-processing. This time scale of seconds allows for a truly
+interactive analysis experience.
 
 5. `climpred` is part of the wider scientific python community, `pangeo`
 [@Eynard:2019]. A wide adoption of `climpred` could standardize prediction model
@@ -141,17 +146,23 @@ time span of the model.
 
 There is seemingly no unified approach to how hindcast initialization dates are aligned
 with observational dates in the academic literature. The authors of `climpred` thus
-identified three techniques, which can be selected by the user: (1) Maximize the degrees
-of freedom by selecting all initialization dates that verify with the available
-observations at each lead. In turn, initializations and verification dates
-are not held constant for each lead. (2) Use the identical set of initializations that
-can verify over the given observational window at all leads. However, the verification
-dates change at each lead. (3) Use the identical verification window at each lead, while
-allowing the set of initializations used at each lead to change. These strategies are
-shown graphically and explained in more detail in the documentation. Note that
-`climpred` offers extensive analysis functionality in addition to forecast verification,
-such as spatiotemporal smoothing [@Goddard:2013], bias removal [@Boer:2016],
-significance testing [@Goddard:2013; @Boer:2016; @DelSole:2016], and a graphics library.
+identified three techniques, which can be selected by the user:
+
+1. Maximize the degrees of freedom by selecting all initialization dates that verify with
+the available observations at each lead. In turn, initializations and verification dates
+are not held constant for each lead.
+
+2. Use the identical set of initializations that can verify over the given observational
+window at all leads. However, the verification dates change at each lead.
+
+(3) Use the identical verification window at each lead, while allowing the set of
+initializations used at each lead to change.
+
+These strategies are shown graphically and explained in more
+detail in the documentation. Note that `climpred` offers extensive analysis functionality
+in addition to forecast verification, such as spatiotemporal smoothing [@Goddard:2013],
+bias removal [@Boer:2016], significance testing [@Goddard:2013; @Boer:2016; @DelSole:2016],
+and a graphics library.
 
 <!-- 58 words -->
 # Use in Academic Literature
