+++ joss/paper.md
@@ -168,9 +168,9 @@ and a graphics library.
 # Use in Academic Literature
 `climpred` has been used to drive analysis in three academic papers so far. @Brady:2020
 used the `HindcastEnsemble` class to highlight multi-year predictability of ocean
-acidification in the California Current; @Spring:2020 used the `PerfectModelEnsemble`
-class to highlight predictability horizons in the global carbon cycle; and
-@Krumhardt:2020 used the `HindcastEnsemble` class to illuminate multi-year
+acidification in the California Current; @Spring:2020 and @Spring:2021 used the 
+`PerfectModelEnsemble` class to highlight predictability horizons in the global carbon
+cycle; and @Krumhardt:2020 used the `HindcastEnsemble` class to illuminate multi-year
 predictability in marine Net Primary Productivity.
 
 # Acknolwedgements
