+++ joss/paper.bib
@@ -357,6 +357,18 @@ key = ""
   number = {9},
 }
 
+
+@article{spring:2021,
+  title = {Trivial Improvements of Predictive Skill Due to Direct Reconstruction of Global Carbon Cycle},
+  author = {Spring, Aaron and Dunkl, Istv{\'a}n and Li, Hongmei and Brovkin, Victor and Ilyina, Tatiana},
+  year = {2021},
+  pages = {1--36},
+  issn = {2190-4979},
+  doi = {10/gh3tn3},
+  journal = {Earth System Dynamics Discussions},
+  language = {English}
+}
+
 @article{Yeager:2018,
   title = {Predicting Near-Term Changes in the {{Earth System}}: {{A}} Large Ensemble of Initialized Decadal Prediction Simulations Using the {{Community Earth System Model}}},
   shorttitle = {Predicting Near-Term Changes in the {{Earth System}}},
