+++ joss/paper.bib
@@ -358,7 +358,7 @@ key = ""
 }
 
 
-@article{spring:2021,
+@article{Spring:2021,
   title = {Trivial Improvements of Predictive Skill Due to Direct Reconstruction of Global Carbon Cycle},
   author = {Spring, Aaron and Dunkl, Istv{\'a}n and Li, Hongmei and Brovkin, Victor and Ilyina, Tatiana},
   year = {2021},
