+++ joss/paper.bib
@@ -69,21 +69,13 @@
   number = {2},
 }
 
-@misc{Eynard:2019,
-type = "Proceedings paper",
-year = "2019",
-title = "The {Pangeo} Big Data Ecosystem and its use at {CNES}",
-journal = "",
-editor = "",
-volume = "",
-number = "",
-pages = "",
-author = "Eynard-Bontemps Guillaume, Abernathey Ryan, Hamman Joseph, Ponte Aurelien, Rath Willi",
-url = "https://archimer.ifremer.fr/doc/00503/61441/",
-organization = "",
-address = "France, USA, Germany",
-abstract = "<p>Pangeo[1] is a community-driven effort for open-source big data initially focused on the Earth System Sciences. One of its primary goals is to enable scientists in analyzing petascale datasets both on classical high-performance computing (HPC) and on public cloud infrastructure. In only a few years, Pangeo has grown into a very productive community collaborating on the development of open-source analysis tools for science. It provides a set of example deployments based on open-source Scientific Python packages like Jupyter[2], Dask[3], and Xarray[4] that bring together scientists and developer with their actual use-cases. In this paper, we first describe Pangeo ecosystem and community. We then present its impact on the work of scientists from CNES on the HPC deployment there. We conclude with a future outlook for Pangeo in this agency and beyond.</p>",
-key = ""
+@article{Abernathey:2017,
+  title = {Pangeo {{NSF Earthcube Proposal}}},
+  author = {Abernathey, Ryan and Paul, Kevin and Hamman, Joe and Rocklin, Matthew and Lepore, Chiara and Tippett, Michael and Henderson, Naomi and Seager, Richard and May, Ryan and Del Vento, Davide},
+  year = {2017},
+  publisher = {{figshare}},
+  doi = {10/gh3ts4},
+  language = {en}
 }
 
 
