+++ joss/paper.bib
@@ -78,6 +78,13 @@
   language = {en}
 }
 
+@misc{Eynard:2019,
+  type = "Proceedings paper",
+  year = "2019",
+  title = "The {Pangeo} Big Data Ecosystem and its use at {CNES}",
+  author = "Eynard-Bontemps Guillaume, Abernathey Ryan, Hamman Joseph, Ponte Aurelien, Rath Willi",
+  url = "https://archimer.ifremer.fr/doc/00503/61441/",
+}
 
 @article{Eyring:2016,
   title = {Overview of the {{Coupled Model Intercomparison Project Phase}} 6 ({{CMIP6}}) Experimental Design and Organization},
