+++ joss/paper.bib
@@ -84,6 +84,8 @@
   title = "The {Pangeo} Big Data Ecosystem and its use at {CNES}",
   author = "Eynard-Bontemps Guillaume, Abernathey Ryan, Hamman Joseph, Ponte Aurelien, Rath Willi",
   url = "https://archimer.ifremer.fr/doc/00503/61441/",
+  doi = "10.2760/848593",
+  isbn = "978-92-76-00034-1",
 }
 
 @article{Eyring:2016,
