### 2021-02-18 c816a39
Apply suggestions from code review
