+++ joss/paper.bib
@@ -82,7 +82,7 @@
   type = "Proceedings paper",
   year = "2019",
   title = "The {Pangeo} Big Data Ecosystem and its use at {CNES}",
-  author = "Eynard-Bontemps Guillaume, Abernathey Ryan, Hamman Joseph, Ponte Aurelien, Rath Willi",
+  author = "Eynard-Bontemps, Guillaume and Abernathey, Ryan and Hamman, Joseph and Ponte, Aurelien and Rath Willi",
   url = "https://archimer.ifremer.fr/doc/00503/61441/",
   doi = "10.2760/848593",
   isbn = "978-92-76-00034-1",
