### 2021-02-18 069cc3f
📝 Update JOSS paper (#553)
* Update paper.md

* Update paper.bib
