### 2021-02-18 #2781 @aaronspring
Hey @dhhagan, we are ready. the final version is now in master: https://github.com/pangeo-data/climpred/tree/master/joss
