### 2021-02-18 c10f690
Update climpred/bias_removal.py
