### 2021-02-18 b529728
Update CHANGELOG.rst
