+++ CHANGELOG.rst
@@ -27,6 +27,7 @@ Bug fixes
 - :py:meth:`~climpred.classes.HindcastEnsemble.remove_bias` now carries attributes.
           (:issue:`531`, :pr:`551`) `Aaron Spring`_.
 
+
 climpred v2.1.1 (2020-10-13)
 ============================
 
