### 2021-02-18 f265afe
remove_bias keeps attrs (#551)
* add note about imports to notebooks

* keep attrs
