### 2021-02-22 #2781 @dhhagan
Hey @aaronspring - the paper looks good! Could you please tag a release and archive and then report the final version number and archive DOI in this thread?
