+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ What's New
 ==========
 
 
-climpred v2.1.2 (2021-xx-xx)
+climpred v2.1.2 (2021-01-22)
 ============================
 
 New Features
@@ -12,13 +12,16 @@ New Features
 - Function to calculate predictability horizon
   :py:func:`~climpred.predictability_horizon.predictability_horizon` based on condition.
   (:issue:`46`, :pr:`521`) `Aaron Spring`_.
+- Fixed version for our Journal of Open Source Software (JOSS) article: See `review
+  <https://github.com/openjournals/joss-reviews/issues/2781>`_
 
 Bug fixes
 ---------
-- :py:meth:`~climpred.classes.PredictionEnsemble.smooth` now carries lead.attrs (:issue: `527`, pr:`521`) `Aaron Spring`_.
+- :py:meth:`~climpred.classes.PredictionEnsemble.smooth` now carries ``lead.attrs``
+  (:issue:`527`, pr:`521`) `Aaron Spring`_.
 - :py:meth:`~climpred.classes.PerfectModelEnsemble.verify` now works with ``references``
-  also for geospatial inputs, which returned ``NaN``s before.
-  (:issue: `522`, pr:`521`) `Aaron Spring`_.
+  also for geospatial inputs, which returned ``NaN`` before.
+  (:issue:`522`, pr:`521`) `Aaron Spring`_.
 - :py:meth:`~climpred.classes.PredictionEnsemble.plot` now shifts composite lead
   frequencies like ``days``, ``pentads``, ``seasons`` correctly.
   (:issue:`532`, :pr:`533`) `Aaron Spring`_.
