+++ joss/paper.bib
@@ -22,7 +22,6 @@
   issn = {2041-1723},
   doi = {10/ggtpks},
   journal = {Nature Communications},
-  language = {en},
   number = {1},
 }
 
@@ -35,7 +34,6 @@
   issn = {0930-7575, 1432-0894},
   doi = {10/gd7hfq},
   journal = {Climate Dynamics},
-  language = {en},
 }
 
 @misc{CDO,
@@ -63,7 +61,6 @@
   issn = {0027-0644},
   doi = {10/f782pf},
   journal = {Monthly Weather Review},
-  language = {en},
   number = {2},
 }
 
@@ -73,7 +70,6 @@
   year = {2017},
   publisher = {{figshare}},
   doi = {10/gh3ts4},
-  language = {en}
 }
 
 @misc{Eynard:2019,
@@ -109,7 +105,6 @@
   issn = {0930-7575, 1432-0894},
   doi = {10/f4jjvf},
   journal = {Climate Dynamics},
-  language = {en},
   number = {1-2},
 }
 
@@ -123,7 +118,6 @@
   issn = {0036-8075, 1095-9203},
   doi = {10/dp65gs},
   journal = Science,
-  language = {en},
   number = {5297},
   pmid = {8985005},
 }
@@ -151,8 +145,6 @@
   issn = {2049-9647},
   doi = {10/gdqdmw},
   journal = {Journal of Open Research Software},
-  keywords = {data analysis,data; data handling,multidimensional,netCDF,pandas,Python},
-  language = {en},
   number = {1},
 }
 
@@ -167,8 +159,6 @@
   issn = {2049-9647},
   doi = {10/gfxr7x},
   journal = {Journal of Open Research Software},
-  keywords = {annotations,CRPSS,Decadal prediction,MiKlip,model evaluation,MSESS,read,skill score,verification},
-  language = {en},
   number = {1},
 }
 
@@ -216,7 +206,6 @@
   issn = {0930-7575, 1432-0894},
   doi = {10/gdsnf8},
   journal = {Climate Dynamics},
-  language = {en},
 }
 
 @article{Krumhardt:2020,
@@ -228,8 +217,6 @@
   issn = {1944-9224},
   doi = {10/gg9ss8},
   journal = {Global Biogeochemical Cycles},
-  keywords = {Large Marine Ecosystems,net primary production,prediction},
-  language = {en},
   number = {6},
 }
 
@@ -255,7 +242,6 @@
   issn = {1364-8152},
   doi = {10/gc9wzk},
   journal = {Environmental Modelling \& Software},
-  keywords = {annotations,Climate forecast,Forecast verification,R package,read,Skill score},
 }
 
 @article{Meehl:2009,
@@ -282,7 +268,6 @@
   issn = {0003-0007},
   doi = {10/ggvcqv},
   journal = {Bulletin of the American Meteorological Society},
-  language = {en},
   number = {6},
 }
 
@@ -309,7 +294,6 @@
   pages = {126--132},
   address = {{Austin, Texas}},
   doi = {10/gfz6s5},
-  language = {en},
 }
 
 @article{Seferian:2018,
@@ -320,7 +304,6 @@
   issn = {00948276},
   doi = {10/gdb424},
   journal = {Geophysical Research Letters},
-  language = {en},
 }
 
 @article{Spring:2020,
@@ -332,7 +315,6 @@
   issn = {1944-8007},
   doi = {10/ggtbv2},
   journal = {Geophysical Research Letters},
-  keywords = {atmospheric CO2,carbon fluxes,decadal predictability,Earth System Model,internal variability},
   number = {9},
 }
 
@@ -345,7 +327,6 @@
   issn = {2190-4979},
[…]
