### 2021-02-22 75f94c3
Release v2.1.2 (#556)
* changelog

* joss v2.1.2
