+++ joss/paper.bib
@@ -7,7 +7,6 @@
   pages = {3751--3777},
   issn = {1991-9603},
   doi = {10/f89qdf},
-  abstract = {The Decadal Climate Prediction Project (DCPP) is a coordinated multi-model investigation into decadal climate prediction, predictability, and variability. The DCPP makes use of past experience in simulating and predicting decadal variability and forced climate change gained from the fifth Coupled Model Intercomparison Project (CMIP5) and elsewhere. It builds on recent improvements in models, in the reanalysis of climate data, in methods of initialization and ensemble generation, and in data treatment and analysis to propose an extended comprehensive decadal prediction investigation as a contribution to CMIP6 (Eyring et al., 2016) and to the WCRP Grand Challenge on Near Term Climate Prediction (Kushnir et al., 2016). The DCPP consists of three components. Component~A comprises the production and analysis of an extensive archive of retrospective forecasts to be used to assess and understand historical decadal prediction skill, as a basis for improvements in all aspects of end-to-end decadal prediction, and as a basis for forecasting on annual to decadal timescales. Component~B undertakes ongoing production, analysis and dissemination of experimental quasi-real-time multi-model forecasts as a basis for potential operational forecast production. Component~C involves the organization and coordination of case studies of particular climate shifts and variations, both natural and naturally forced (e.g. the ``hiatus'', volcanoes), including the study of the mechanisms that determine these behaviours. Groups are invited to participate in as many or as few of the components of the DCPP, each of which are separately prioritized, as are of interest to them.The Decadal Climate Prediction Project addresses a range of scientific issues involving the ability of the climate system to be predicted on annual to decadal timescales, the skill that is currently and potentially available, the mechanisms involved in long timescale variability, and the production of forecasts of benefit to both science and society.},
   journal = {Geosci. Model Dev.},
   number = {10},
 }
@@ -22,9 +21,7 @@
   publisher = {{Nature Publishing Group}},
   issn = {2041-1723},
   doi = {10/ggtpks},
-  abstract = {Threats to marine ecosystems are increasing due to ocean acidification, but trends are spatiotemporally difficult to monitor or predict. Here the authors use an Earth system model to accurately predict surface pH changes in the economically and ecologically important California Current System.},
   journal = {Nature Communications},
-  language = {en},
   number = {1},
 }
 
@@ -37,7 +34,6 @@
   issn = {0930-7575, 1432-0894},
   doi = {10/gd7hfq},
   journal = {Climate Dynamics},
-  language = {en},
 }
 
 @misc{CDO,
@@ -65,7 +61,6 @@
   issn = {0027-0644},
   doi = {10/f782pf},
   journal = {Monthly Weather Review},
-  language = {en},
   number = {2},
 }
 
@@ -75,7 +70,6 @@
   year = {2017},
   publisher = {{figshare}},
   doi = {10/gh3ts4},
-  language = {en}
 }
 
 @misc{Eynard:2019,
@@ -97,7 +91,6 @@
   pages = {1937--1958},
   issn = {1991-9603},
   doi = {10/gbpwpt},
[…]
