+++ README.rst
@@ -28,6 +28,9 @@ Verification of weather and climate forecasts.
 
 .. |joss| image:: https://joss.theoj.org/papers/246d440e3fcb19025a3b0e56e1af54ef/status.svg
     :target: https://joss.theoj.org/papers/246d440e3fcb19025a3b0e56e1af54ef
+    
+.. |doi| image:: https://zenodo.org/badge/DOI/10.5281/zenodo.4556086.svg
+   :target: https://doi.org/10.5281/zenodo.4556086
 
 .. |ci|  image:: https://github.com/pangeo-data/climpred/workflows/climpred%20testing/badge.svg
 
