### 2021-02-23 2fd3661
allow multiple types for dim
