### 2021-02-23 f0d6651
rm coords fix in conftest
