+++ .github/workflows/climpred_testing.yml
@@ -40,6 +40,37 @@ jobs:
           file: ./coverage.xml
           fail_ci_if_error: false
 
+  doctest:  # tests all docstrings
+    name: Doctests
+    runs-on: "ubuntu-latest"
+    defaults:
+      run:
+        shell: bash -l {0}
+    steps:
+      - uses: actions/checkout@v2
+      - uses: conda-incubator/setup-miniconda@v2
+        with:
+          channels: conda-forge
+          channel-priority: strict
+          mamba-version: "*"
+          activate-environment: climpred-minimum-tests
+          auto-update-conda: false
+          python-version: "3.8"
+
+      - name: Install conda dependencies
+        run: |
+          mamba env update -f ci/requirements/minimum-tests.yml
+      - name: Install climpred
+        run: |
+          python -m pip install --no-deps -e .
+      - name: Conda info
+        run: conda info
+      - name: Conda list
+        run: conda list
+      - name: Run doctests
+        run: |
+          python -m pytest --doctest-modules climpred --ignore climpred/tests
+
   docs_notebooks:  # Checks that pre-compiled notebooks in docs still work.
     runs-on: ubuntu-latest
     defaults:
