### 2021-02-23 7852f8f
allow multiple types for dim (#558)
* allow multiple types for dim

* refactor

* rm coords fix in conftest
