### 2021-02-23 1c86fd5
GHA: Cancel previous runs in testing and install
