+++ .github/workflows/climpred_testing.yml
@@ -14,6 +14,10 @@ jobs:
       matrix:
         python-version: [3.6, 3.7, 3.8, 3.9]
     steps:
+      - name: Cancel previous runs
+        uses: styfle/cancel-workflow-action@0.6.0
+        with:
+          access_token: ${{ github.token }}
       - uses: actions/checkout@v2
       - name: Set up conda
         uses: conda-incubator/setup-miniconda@v2
@@ -47,6 +51,10 @@ jobs:
       run:
         shell: bash -l {0}
     steps:
+      - name: Cancel previous runs
+        uses: styfle/cancel-workflow-action@0.6.0
+        with:
+          access_token: ${{ github.token }}
       - uses: actions/checkout@v2
       - uses: conda-incubator/setup-miniconda@v2
         with:
@@ -77,6 +85,10 @@ jobs:
         run:
           shell: bash -l {0}
     steps:
+      - name: Cancel previous runs
+        uses: styfle/cancel-workflow-action@0.6.0
+        with:
+          access_token: ${{ github.token }}
       - uses: actions/checkout@v2
       - name: Set up Conda
         uses: conda-incubator/setup-miniconda@v2
