+++ climpred/metrics.py
@@ -60,10 +60,9 @@ def _get_norm_factor(comparison):
 
     Example:
         >>> # check skill saturation value of roughly 1 for different comparisons
-        >>> metric = 'nrmse'
         >>> for c in ['m2m', 'm2e', 'm2c', 'e2c']:
-                s = compute_perfect_model(ds, control, metric=metric, comparison=c)
-                s.plot(label=' '.join([metric,c]))
+        ...        s = perfect_model.verify(metric='nrmse', dim=None, comparison=c)
+        ...        s.tos.plot(label=' '.join(['nrmse',c]))
         >>> plt.legend()
 
     Reference:
@@ -1736,20 +1735,20 @@ def _brier_score(forecast, verif, dim=None, **metric_kwargs):
 
         Option 1. Pass with keyword `logical`: (Works also for PerfectModelEnsemble)
 
-        >>> hindcast.verify(metric='brier_score', comparison='m2o', \
-                dim='member', alignment='same_verifs', logical=pos)
+        >>> hindcast.verify(metric='brier_score', comparison='m2o',
+        ...     dim='member', alignment='same_verifs', logical=pos)
 
         Option 2. Pre-process to generate a binary forecast and verification product:
 
-        >>> hindcast.map(pos).verify(metric='brier_score', \
-                comparison='m2o', dim='member', alignment='same_verifs')
+        >>> hindcast.map(pos).verify(metric='brier_score',
+        ...     comparison='m2o', dim='member', alignment='same_verifs')
 
         Option 3. Pre-process to generate a probability forecast and binary
         verification product. Because `member` no present in `hindcast`, use
         ``comparison='e2o'`` and ``dim=[]``:
 
-        >>> hindcast.map(pos).mean('member').verify(metric='brier_score', \
-                comparison='e2o', dim=[], alignment='same_verifs')
+        >>> hindcast.map(pos).mean('member').verify(metric='brier_score',
+        ...     comparison='e2o', dim=[], alignment='same_verifs')
     """
     forecast, verif, metric_kwargs, dim = _extract_and_apply_logical(
         forecast, verif, metric_kwargs, dim
@@ -1815,15 +1814,15 @@ def _threshold_brier_score(forecast, verif, dim=None, **metric_kwargs):
         * :py:func:`~xskillscore.threshold_brier_score`
 
     Example:
-        >>> import climpred
+
         >>> init = climpred.tutorial.load_dataset("CESM-DP-SST")
         >>> obs = climpred.tutorial.load_dataset("ERSST")
         >>> obs = obs - obs.mean('time')
         >>> hindcast = climpred.HindcastEnsemble(init).add_observations(obs)
 
         >>> # get threshold brier score for each init
-        >>> hindcast.verify(metric='threshold_brier_score', comparison='m2o', \
-                dim='member', threshold=.2, alignment='same_verifs')
+        >>> hindcast.verify(metric='threshold_brier_score', comparison='m2o',
+        ...     dim='member', threshold=.2, alignment='same_verifs')
         <xarray.Dataset>
         Dimensions:  (init: 52, lead: 10)
         Coordinates:
@@ -1834,8 +1833,8 @@ def _threshold_brier_score(forecast, verif, dim=None, **metric_kwargs):
             SST      (lead, init) float64 0.0 0.0 0.0 0.0 0.0 ... 1.0 0.01 0.0 0.0 0.0
 
         >>> # multiple thresholds averaging over init dimension
-        >>> hindcast.verify(metric='threshold_brier_score', comparison='m2o', \
-                dim=['member', 'init'], threshold=[.2, .3], alignment='same_verifs')
[…]

[…]
