### 2021-02-24 4703639
fix dim=None expansion for e2c comparison
