### 2021-02-24 b0c595f
skip smooth doctests with xesmf
