### 2021-02-24 d118250
condense conftest docnamespace
