### 2021-02-24 7c17581
test contingency.table
