+++ climpred/metrics.py
@@ -60,10 +60,10 @@ def _get_norm_factor(comparison):
 
     Example:
         >>> # check skill saturation value of roughly 1 for different comparisons
-        >>> for c in ['m2m', 'm2e', 'm2c', 'e2c']:
+        >>> for c in ['m2m', 'm2e', 'm2c', 'e2c']:  # doctest: +ELLIPSIS
         ...        s = perfect_model.verify(metric='nrmse', dim=None, comparison=c)
-        ...        s.tos.plot(label=' '.join(['nrmse',c]))
-        >>> plt.legend()
+        ...        s.tos.plot(label='nrmse {c}')
+        [...
 
     Reference:
         * Séférian, Roland, Sarah Berthet, and Matthieu Chevallier. “Assessing
@@ -136,7 +136,6 @@ def _extract_and_apply_logical(forecast, verif, metric_kwargs, dim):
                 f"Expected dimension `member` in forecast, found {list(forecast.dims)}"
             )
         # rename dim to time if forecast and verif dims allow
-
         return forecast, verif, metric_kwargs, dim
     elif (
         comparison.name == "e2o"
@@ -1736,19 +1735,41 @@ def _brier_score(forecast, verif, dim=None, **metric_kwargs):
         Option 1. Pass with keyword `logical`: (Works also for PerfectModelEnsemble)
 
         >>> hindcast.verify(metric='brier_score', comparison='m2o',
-        ...     dim='member', alignment='same_verifs', logical=pos)
+        ...     dim=['member', 'init'], alignment='same_verifs', logical=pos)
+        <xarray.Dataset>
+        Dimensions:  (lead: 10)
+        Coordinates:
+          * lead     (lead) int32 1 2 3 4 5 6 7 8 9 10
+            skill    <U11 'initialized'
+        Data variables:
+            SST      (lead) float64 0.115 0.1121 0.1363 0.125 ... 0.1654 0.1675 0.1873
 
         Option 2. Pre-process to generate a binary forecast and verification product:
 
+        >>> # TODO: should be identical to above
         >>> hindcast.map(pos).verify(metric='brier_score',
-        ...     comparison='m2o', dim='member', alignment='same_verifs')
+        ...     comparison='m2o', dim=['member', 'init'], alignment='same_verifs')
+        <xarray.Dataset>
+        Dimensions:  (lead: 10)
+        Coordinates:
+          * lead     (lead) int32 1 2 3 4 5 6 7 8 9 10
+            skill    <U11 'initialized'
+        Data variables:
+            SST      (lead) float64 0.1577 0.175 0.1904 0.1654 ... 0.2038 0.2135 0.2269
 
         Option 3. Pre-process to generate a probability forecast and binary
         verification product. Because `member` no present in `hindcast`, use
-        ``comparison='e2o'`` and ``dim=[]``:
+        ``comparison='e2o'`` and ``dim='init'``:
 
         >>> hindcast.map(pos).mean('member').verify(metric='brier_score',
-        ...     comparison='e2o', dim=[], alignment='same_verifs')
+        ...     comparison='e2o', dim='init', alignment='same_verifs')
+        <xarray.Dataset>
+        Dimensions:  (lead: 10)
+        Coordinates:
+          * lead     (lead) int32 1 2 3 4 5 6 7 8 9 10
+            skill    <U11 'initialized'
+        Data variables:
+            SST      (lead) float64 0.115 0.1121 0.1363 0.125 ... 0.1654 0.1675 0.1873
     """
     forecast, verif, metric_kwargs, dim = _extract_and_apply_logical(
         forecast, verif, metric_kwargs, dim
@@ -1928,7 +1949,7 @@ def _crps(forecast, verif, dim=None, **metric_kwargs):
[…]
