### 2021-02-24 b1548b1
dont add attrs in pm.verify
