### 2021-02-24 a25f9bd
refactor member mean in reliability and discrimination
