### 2021-02-24 f294f92
smooth doctest
