### 2021-02-24 cc84207
fix smooth doctest
