+++ climpred/classes.py
@@ -405,22 +405,26 @@ class PredictionEnsemble:
         datasets = self._datasets.copy()
 
         # More explicit than nested dictionary comprehension.
-        for outer_k, outer_v in datasets.items():
-            # If initialized, control, uninitialized and just a singular
-            # dataset, apply the function directly to it.
-            if isinstance(outer_v, xr.Dataset):
-                datasets.update({outer_k: func(outer_v, *args, **kwargs)})
-            else:
-                # If a nested dictionary is encountered (i.e., a set of
-                # observations) apply to each individually.
-                #
-                # Similar to the ``add_observations`` method, this only seems to
-                # avoid inplace operations by copying the nested dictionary
-                # separately and then updating the main dictionary.
-                temporary_dataset = self._datasets[outer_k].copy()
-                for inner_k, inner_v in temporary_dataset.items():
-                    temporary_dataset.update({inner_k: func(inner_v, *args, **kwargs)})
-                datasets.update({outer_k: temporary_dataset})
+        for key, ds in datasets.items():
+            # If ds is xr.Dataset, apply the function directly to it. else, e.g. for {} ignore
+            if isinstance(ds, xr.Dataset):
+                dim = kwargs.get("dim", "")
+                if "_or_" in dim:
+                    dims = dim.split("_or_")
+                    if set(dims).issubset(ds.dims):
+                        raise ValueError(
+                            f"{dims} cannot be both in {key} dataset, found {ds.dims}"
+                        )
+                    kwargs_dim0 = kwargs.copy()
+                    kwargs_dim0["dim"] = dims[0]
+                    kwargs_dim1 = kwargs.copy()
+                    kwargs_dim1["dim"] = dims[1]
+                    if dims[0] in ds.dims and dims[1] not in ds.dims:
+                        datasets.update({key: func(ds, *args, **kwargs_dim0)})
+                    if dims[1] in ds.dims and dims[0] not in ds.dims:
+                        datasets.update({key: func(ds, *args, **kwargs_dim1)})
+                else:
+                    datasets.update({key: func(ds, *args, **kwargs)})
         # Instantiates new object with the modified datasets.
         return self._construct_direct(datasets, kind=self.kind)
 
