+++ climpred/tests/test_map.py
@@ -32,11 +32,13 @@ def test_PredictionEnsemble_map_dim_or(hindcast_hist_obs_1d):
     """Tests that PredictionEnsemble allows dim0_or_dim1 as kwargs without UserWarning."""
     with pytest.warns(None):  # no warnings
         he_or = hindcast_hist_obs_1d.map(rm_poly, dim="init_or_time", order=2)
+    assert he_or != hindcast_hist_obs_1d
 
     with pytest.warns(UserWarning) as record:  # triggers warnings
         he_chained = hindcast_hist_obs_1d.map(rm_poly, dim="init", order=2).map(
             rm_poly, dim="time", order=2
         )
+    assert he_chained != hindcast_hist_obs_1d
 
     assert len(record) == 3  # for init, uninit and obs
     for r in record:
