### 2021-02-27 e06242c
⚠️ map functions and xarray calls show warnings (#552)
* user warnings for errors

* add tests

* nicer warnings

* rerun selected notebooks with warnings

* rerun decadal examples no filter warnings, rm cartopy

* allow dim0_or_dim1

* rerun notebooks
