### 2021-03-01 #2781 @kthyng
Hi @aaronspring! Can you update the metadata on your Zenodo archive so that the title and author list exactly match your JOSS paper?
