### 2021-03-01 #2781 @aaronspring
> Hi @aaronspring! Can you update the metadata on your Zenodo archive so that the title and author list exactly match your JOSS paper?

I can only add one institution, hope thats fine. https://zenodo.org/record/4556086
