### 2021-03-01 87ac535
Update paper.md (#557)
* Update index.rst

* Update paper.md
