### 2021-03-01 #2781 @aaronspring
> @aaronspring Also please correct your spelling of "Acknowledgements" in the paper. Then we're ready to move forward.

thanks. now fixed in `master`: https://github.com/pangeo-data/climpred/blob/master/joss/paper.md
