+++ climpred/metrics.py
@@ -128,9 +128,9 @@ def _extract_and_apply_logical(forecast, verif, metric_kwargs, dim):
         if not callable(logical):
             raise ValueError(f"`logical` must be `callable`, found {type(logical)}")
         dim = _remove_member_from_dim_or_raise(dim)
-        if "member" in forecast.dims:  # apply logical function to get
-            forecast = logical(forecast).mean("member")  # forecast probability
-            verif = logical(verif)  # binary outcome
+        # apply logical function to get forecast probability
+        forecast = logical(forecast).mean("member")
+        verif = logical(verif)  # binary outcome
         else:
             raise ValueError(
                 f"Expected dimension `member` in forecast, found {list(forecast.dims)}"
