+++ climpred/metrics.py
@@ -1779,7 +1779,7 @@ def _brier_score(forecast, verif, dim=None, **metric_kwargs):
     forecast, verif, metric_kwargs, dim = _extract_and_apply_logical(
         forecast, verif, metric_kwargs, dim
     )
-    if 'member' in foreforecast.dims:  # rm this with xs 0.0.19
+    if 'member' in forecast.dims:  # rm this with xs 0.0.19
         forecast = forecast.mean('member')
         dim = dim.copy()
         dim.remove('member')
